==> scripts/R/02_update_metadata.Rmd <==
---
title: "checking files"
author: "Kimberly Olney, Ph.D"
date: "07/13/2022"
output:
  html_document:
    theme: cerulean
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---

# Setup
```{r setup}
knitr::opts_knit$set(root.dir = ".")
```
# User defined variables
```{r set_variables}
pathToRef = c("/research/labs/neurology/fryer/projects/references/human/")
pathToRawData = c("/research/labs/neurology/fryer/projects/LBD_CWOW/")
```
# Read data
```{r read_data}
# read in metadata
metadata <- read.delim(paste0(pathToRawData, "metadata.tsv"), 
                       header = TRUE,
                       sep = "\t")
WGS <- read.delim(paste0(pathToRawData, "WGS_sampleID.txt"), 
                       header = TRUE,
                       sep = "\t")
RIN <- read.delim(paste0(pathToRawData, "RIN.tsv"), 
                       header = TRUE,
                       sep = "\t")
Sex_inf <- read.delim("sex_check_RNA.txt", 
                      header = TRUE, 
                      sep = "\t")
snakemake_sample_info <- read.delim("../snakemake/sample_read_group_info.txt",
                                    header = FALSE)
```
# Add columns to metadata
```{r}
# rename sample column to NPID for merging with metadata
names(RIN)[names(RIN) == "Sample"] <- "NPID"
RIN_metadata <- merge(metadata, RIN, by = "NPID", all.x = TRUE)
WGS_metadata <- merge(RIN_metadata, WGS, by = "NPID", all.x = TRUE)

# drop columns and rename
names(Sex_inf)[names(Sex_inf) == "reported"] <- "sex_inferred"
Sex_inf$observed <- NULL
Sex_inf$sex_check <- NULL
unique_sex_inf <- unique(Sex_inf)
# Add sex inference to metadata
sex_metadata <-
  merge(WGS_metadata, unique_sex_inf, by = "NPID", all.x = TRUE)

# get the sampleID from the counts file to match with metadata
snakemake_sample_info[c('NPID', 'flowcell')] <-
  str_split_fixed(snakemake_sample_info$V1, '.FC', 2)
snakemake_sample_info[c('flowcell_and_lane', 'lane')] <-
  str_split_fixed(snakemake_sample_info$flowcell, '_R1', 2)
# drop columns
snakemake_sample_info$V1 <- NULL
snakemake_sample_info$V2 <- NULL
snakemake_sample_info$flowcell <- NULL
snakemake_sample_info$lane <- NULL
# merge with sex_metadata
snake_metadata <-
  merge(
    sex_metadata,
    snakemake_sample_info,
    by = "NPID",
    all.x = TRUE,
    all.y = TRUE
  )

snake_metadata$RNA.config.ID <-
  paste0(snake_metadata$NPID, "_FC", snake_metadata$flowcell_and_lane)
write.table(
  snake_metadata,
  paste0(pathToRawData, "metadata_expanded.tsv"),
  quote = FALSE,
  row.names = FALSE,
  sep = "\t"
)
```
