==> scripts/R/03_variance.Rmd <==
---
title: "Variance in expression data LBD samples"
author: "Kimberly Olney, Ph.D."
date: "09/30/2022"
output:
  html_document:
    df_print: paged
  pdf_document: default
params:
  args: myarg
---
# setup
```{r setup}
knitr::opts_knit$set(root.dir = ".")
```

```{r libraries, message=FALSE, warning=FALSE}
library('variancePartition')
library('edgeR')
library('BiocParallel')
library('Hmisc')
```

# user defined variables
```{r set_variables}
source(here::here("scripts/R", "file_paths_and_colours.R"))
directory <- c("pass1/")
condition <- c("All_samples")

myContrasts <- c("LBDvsControl = LBD - CONTROL",
  "LBDvsAD = LBD - AD",
  "LBDvsPA = LBD - PA",
  "ADvsControl = AD - CONTROL", 
  "PAvsControl = PA - CONTROL",
  "ADvsPA = AD - PA")
tool = c("star")
typeOfCount <-  c(".bamReadsPerGene.out.tab")
```

# read in DGE object & metadata
```{r}
dge <- readRDS(paste0("../../rObjects/", directory, "LBD_SCC_dge.raw.rds"))
metadata <- dge$samples
dge.filtered.norm <- readRDS(paste0("../../rObjects/", condition, ".dge.filtered.norm.rds"))
# some samples are missing RIN values. 
# Replace NA with 1 so as to be able to account for RIN in the model. 
# This is necessary to be able include RIN as a covariate in voom
dge.filtered.norm$samples$RIN[is.na(dge.filtered.norm$samples$RIN)] <- 1
info <- as.data.frame(dge.filtered.norm$samples)
genes <- dge.filtered.norm$genes
```

# make contrast plots
```{r}
design <- model.matrix(~ 0 + TYPE + RIN + Age + sex_inferred + APOE + Race, dge.filtered.norm$samples)

form <- ~ 0 + TYPE
myDreamContrasts <- c(LBDvsControl = "TYPELBD - TYPECONTROL",
  LBDvsAD = "TYPELBD - TYPEAD",
  LBDvsPA = "TYPELBD - TYPEPA",
  ADvsControl = "TYPEAD - TYPECONTROL", 
  PAvsControl = "TYPEPA - TYPECONTROL",
  ADvsPA = "TYPEAD - TYPEPA")
myDreamContrasts<- rev(myDreamContrasts)

L = makeContrastsDream(form, dge$samples, contrasts = myDreamContrasts)
colnames(L) <- gsub("TYPE","",colnames(L))
rownames(L) <- gsub("TYPE","", rownames(L))


# Visualize contrast matrix
plotContrasts(L) 
```

# CCA
```{r}
form <- ~ MF.NFT + 
MF.Amyloid + 
MF.SP + 
MF.Tau + 
MF.LB +
MF.NFT +
Brain.wt + 
Duration + 
Age + 
PMI + 
RIN + 
Braak.NFT + 
Thal.amyloid + 
Cing.LB + 
Cing.Synuclein + 
APOE +
LBD.type +
AD.subtype + 
CDLB + 
TDP.type +
VaD +
MAPT + 
GRN + 
Race +
TMEM106b +
Race + 
FHx +
sex_inferred 

C = canCorPairs(form, info)
#plotCorrMatrix( C )

cor.mtest <- function(C, ...) {
    C <- as.matrix(C)
    n <- ncol(C)
    p.mat<- matrix(NA, n, n)
    diag(p.mat) <- 0
    for (i in 1:(n - 1)) {
        for (j in (i + 1):n) {
            tmp <- cor.test(C[, i], C[, j], ...)
            p.mat[i, j] <- p.mat[j, i] <- tmp$p.value
        }
    }
  colnames(p.mat) <- rownames(p.mat) <- colnames(C)
  p.mat
}
# matrix of the p-value of the correlation
p.mat <- cor.mtest(C)
col <- colorRampPalette(c("#4477AA", "#77AADD", "#FFFFFF", "#EE9988", "#BB4444"))
  corrplot(C, method="color", col=col(200),  
         type="upper", order="hclust", 
         addCoef.col = "black", # Add coefficient of correlation
         tl.col="black", tl.srt=45, #Text label color and rotation
         diag=FALSE, col.lim = c(0, 1)
         )
  
path <- paste0("../../results/", tool ,"/varpart/LBD_gene_CCA")
saveToPDF(paste0(path, ".pdf"), width = 16, height = 16)
```

# fit variance 
```{r}
design <- model.matrix(~ 0 + TYPE + RIN + Age + sex_inferred + APOE + Race, dge.filtered.norm$samples)
v <- voom(dge.filtered.norm, design, plot=TRUE)
voomCounts <- v$E
voomCountsMatrix <- data.matrix(voomCounts, rownames.force = NA)
```

```{r, eval=FALSE}
form <- ~ (1|TYPE) + 
  (1|TDP.43) +
  (1|sex_inferred) + 
  (1|VaD) + 
  (1|Race) + 
  (1|APOE) + 
  (1|FHx) + 
  Age + 
  MF.SP + 
  MF.NFT + 
  MF.Amyloid +
  RIN + 
  PMI + 
  Braak.NFT + 
  Brain.wt + 
  Duration

# fit model and extract variance percents
varPart <- fitExtractVarPartModel(voomCountsMatrix, form, info, 
                                  showWarnings=FALSE)
plotVarPart(sortCols(varPart), label.angle=60 )

path <- paste0("../../results/", tool, "/varpart/varpart")
saveToPDF(paste0(path, ".pdf"), width = 8.2, height = 5.2)

varPart$gene_id <- rownames(varPart)
variance_explained <- merge(varPart, genes, by = "gene_id")
write.table(variance_explained, 
            paste0("../../results/", tool ,
            "/varpart/variance_explained.tsv"), 
            sep = "\t", 
            quote = FALSE)
```

# re-examine CCA with the variables in the varpart model 
```{r, eval = FALSE}
form <- ~ TYPE + 
  sex_inferred + 
  VaD + 
  Race + 
  APOE + 
  FHx + 
  Age + 
  MF.SP + 
  MF.NFT + 
  MF.Amyloid +
  RIN + 
  Braak.NFT + 
  Brain.wt 
C = canCorPairs( form, info)
# Plot correlation matrix
plotCorrMatrix( C )
path <- paste0("../../results/", tool ,"/varpart/CCA_limited_form")
saveToPDF(paste0(path, ".pdf"), width = 16, height = 16)
```
# re-run variance partition with metadata variables that most samples have information for. 
3 samples are missing MF.SP values
4 samples are missing MF.NFT values 
4 samples are missing MF.Amyloid values 
1 sample is missing Braak.NFT
1 samples is missing brain weight 
265 samples are missing PMI information 
288 samples missing duration information 
52 samples are missing information for TDP.43 status
1 sample is missing VaD information 
1 sample is missing FHx information

will re-run excluding PMI, Duration, and TDP.43 status so as to include these samples that otherwise would be missed. 
Will also remove MF.Amyloid as this is highly correlated with MF.SP 
```{r}
form <- ~ (1|TYPE) + 
  (1|sex_inferred) + 
  (1|VaD) + 
  (1|Race) + 
  (1|APOE) + 
  (1|FHx) + 
  Age + 
  MF.SP + 
  MF.NFT + 
  RIN + 
  Braak.NFT + 
  Brain.wt 

# fit model and extract variance percents
varPart <- fitExtractVarPartModel(voomCountsMatrix, form, info, 
                                  showWarnings=FALSE )
plotVarPart(sortCols(varPart), label.angle=60 )

path <- paste0("../../results/", tool, "/varpart/varpart_NA_variables_excluded")
saveToPDF(paste0(path, ".pdf"), width = 8.2, height = 5.2)
varPart$gene_id <- rownames(varPart)
variance_explained <- merge(varPart, genes, by = "gene_id")
write.table(variance_explained, 
            paste0("../../results/", tool ,
            "/varpart/variance_explained_NA_varaibles_excluded.tsv"), 
            sep = "\t", 
            quote = FALSE)
```

```{r}
variance_explained <- read.delim(paste0("../../results/",
                                        tool,
                                        "/varpart/variance_explained_NA_varaibles_excluded.tsv"),
                                        sep = "\t")
# replace gene_id with gene_name
varPart$gene_id <- NULL
row.names(varPart) <-
  make.names(variance_explained$gene_name, unique = TRUE)
row.names(voomCountsMatrix) <-
  make.names(genes$gene_name, unique = TRUE)

vp <- sortCols(varPart)
plotPercentBars(vp[1:10, ])
plotVarPart(vp)
```
# remove technical variability before fitting model
Subtracting the RIN effect can reduce the total variation in the data, so that the contribution of other variables become clearer.

run the residuals function inside the call to fitVarPartModel to avoid storing the large intermediate results.

Voom - Transform count data to log2-counts per million (logCPM), estimate the mean-variance relationship and use this to compute appropriate observational-level weights. The data are then ready for linear modelling.
```{r}
# subtract out effect of: 
design <- model.matrix(~ 0 + TYPE + RIN + Age + sex_inferred + APOE + Race, dge.filtered.norm$samples)
v <- voom(dge.filtered.norm, design, plot=TRUE)
# Voom transformed counts
voomCounts <- v$E
voomCountsMatrix <- data.matrix(voomCounts, rownames.force = NA)

# extract residuals directly without storing intermediate results
# variancePartition seamlessly deals with the result of voom() 
# by default, it seamlessly models the precision weights
# This can be turned off with useWeights=FALSE
formula <- ~ (1|sex_inferred) + 
  (1|Race) + 
  (1|APOE) + 
  Age + 
  RIN 

residList <- fitVarPartModel(v, formula, dge.filtered.norm$samples,
        fxn=residuals )

# convert list to matrix
residMatrix = do.call(rbind, residList)

dge.filtered.norm$samples$VaD <- as.factor(dge.filtered.norm$samples$VaD)
# fit model on residuals
form <- ~ (1|TYPE) + 
  (1|VaD) + 
  (1|FHx) + 
  MF.SP + 
  MF.NFT + 
  Braak.NFT + 
  Brain.wt 

# fit model and extract variance percents
varPartResid <- fitExtractVarPartModel(residMatrix, 
                                       form, 
                                       dge.filtered.norm$samples,
                                       showWarnings=FALSE)
plotVarPart(sortCols(varPartResid), label.angle=60)
varPartResid$gene_id <- NULL
varPartResid$gene_id <- rownames(varPartResid)
variance_explained_correcting_vars <- merge(varPartResid, genes, by = "gene_id")
write.table(variance_explained_correcting_vars, 
            paste0("../../results/", tool ,
            "/varpart/variance_explained_correcting_vars.tsv"), 
            sep = "\t", 
            quote = FALSE)
```


```{r}
colnames(design) <- c(CONTROL, AD, PA, LBD, "RIN", "Age", "sex", "E2E4", "E3E3", "E3E4", "E4E4", "Asian_American", "Caucasian", "Eastern_Indian", "Hispanic", "Native_American")

# fits linear model for each gene given a series of arrays
fit <- lmFit(v, design)

# contrast design for differential expression
contrasts <- makeContrasts(
  LBDvsControl = LBD - CONTROL,
  LBDvsAD = LBD - AD,
  LBDvsPA = LBD - PA,
  ADvsControl = AD - CONTROL, 
  PAvsControl = PA - CONTROL,
  ADvsPA = AD - PA, 
  levels = colnames(design))
head(contrasts)

# save contrast names
allComparisons <- colnames(contrasts)
allComparisons # check

# run contrast analysis
vfit <- contrasts.fit(fit, contrasts = contrasts)

# Compute differential expression based on the empirical Bayes moderation of the
# standard errors towards a common value.
veBayesFit <- eBayes(vfit)
plotSA(veBayesFit, main = "Final Model: Mean-variance Trend")

# save
path <- paste0("../../results/", tool, "/voom/LBD_gene_final_mean_var")
saveToPDF(paste0(path, ".pdf"), width = 6, height = 4)
```

# voom MDS Plot
```{r MDS_voom}
group_colors <- c(control_color, 
                  AD_color, 
                  PA_color, 
                  LBD_color)[v$targets$TYPE]
Sex <- v$targets$Sex

plotMDS(
  v, 
  top = 100, 
  labels = Sex,
  cex = 1, 
  dim.plot = c(1,2), 
  plot = TRUE, 
  col = group_colors, 
  gene.selection = "common"
)

title(expression('Top 100 Genes - Voom (Log'[2]~'CPM)'))

# save
# save
path <- paste0("../../results/", tool, "/MDS/LBD_gene_MDS_voom_sex")
saveToPDF(paste0(path, ".pdf"), width = 5, height = 5)
```

# PCA
```{r}
# Setting the N of genes to use
ntop = length(dge.filtered.norm$genes$gene_id)
# Sorting by the coefficient of variance
means <- rowMeans(voomCountsMatrix)
Pvars <- rowVars(voomCountsMatrix)
cv2 <- Pvars / means ^ 2
select <-
  order(cv2, decreasing = TRUE)[seq_len(min(ntop, length(cv2)))]
head(select)

highly_variable_exp <- ((voomCountsMatrix)[select,])
dim(highly_variable_exp)
# Running PCA
pca_exp <- prcomp(t(highly_variable_exp), scale = F, center = T)
# scale a logical value indicating whether the variables should be scaled
# to have unit variance before the analysis takes place.
# a logical value indicating whether the variables should be shifted to be zero centered.
head(pca_exp$x)[, 1:3]
summary(pca_exp)
# Dataframe with the first 10 PCs
dim1_10 <- data.frame(pca_exp$x[, 1:10])
# Adding metadata
pcaWithMetadata <- merge(dim1_10, dge.filtered.norm$samples, by = 0, all = TRUE)
pcaWithMetadata$group <- factor(pcaWithMetadata$TYPE,
                                 levels = c(CONTROL, AD, PA, LBD, NA))
# Plotting
ggplot(data = pcaWithMetadata, aes(
  x = PC1,
  y = PC2,
  shape = group,
  color = group
)) +
  geom_point(size = 2.5) +
  theme_bw() +
#  xlim(-150, 150) +
#  ylim(-150, 150) +
  scale_color_manual(values = c(control_color, AD_color,
                                PA_color, LBD_color,
                                "azure3")) +
  scale_shape_manual(values = c(control_shape, AD_shape,
                                PA_shape, LBD_shape)) +
  theme(
    plot.title = element_text(size = 12, face = "bold"),
    legend.title = element_text(size = 12),
    legend.text = element_text(size = 10),
    axis.text.x = element_text(size = 10),
    axis.text.y = element_text(size = 10),
    axis.title.x = element_text(size = 10),
    axis.title.y = element_text(size = 10)
  ) +
  theme(legend.title = element_blank()) +
  xlab("PC1 ") +
  ylab("PC2 ")
```
# CCA with PCA 1-4
```{r}
form <- ~ 
  PC1 +
  PC2 +
  PC3 +
  PC4 +
  TYPE + 
  sex_inferred + 
  VaD + 
  Race + 
  APOE + 
  FHx + 
  Age + 
  MF.SP + 
  MF.NFT + 
  MF.Amyloid +
  RIN + 
  Braak.NFT + 
  Brain.wt 
C = canCorPairs(form, pcaWithMetadata)
# Plot correlation matrix
plotCorrMatrix( C )
```

```{r}
saveRDS(veBayesFit, file = paste0("../../rObjects/", condition,".veBayesFit.rds"))
saveRDS(voomCountsMatrix, file = paste0("../../rObjects/", condition,".voomCountsMatrix.rds"))

veBayesFit <- readRDS(paste0("../../rObjects/", condition,".veBayesFit.rds"))
voomCountsMatrix <- readRDS(paste0("../../rObjects/", condition,".voomCountsMatrix.rds"))
```

# number of DEGs
Identify number of differential expressed genes.
```{r}
pval <- 0.05

sumTable <- 
  summary(decideTests(
    vfit,  # object
    adjust.method = "BH", # by default the method = "separate"
    p.value = pval,
    lfc = 0  # numeric, minimum absolute log2-fold change required
  ))

print(paste0(" FDRq < ", pval))
sumTable
topTable(veBayesFit, coef = "ADvsPA")
```
# output DEG tables
```{r}
coef <- 1

for (i in allComparisons) {
  # p < 1, log2fc > 0 
  vTopTableAll <-
    topTable(
      veBayesFit, 
      coef = coef,  
      n = Inf, 
      p.value = 1,
      lfc = 0 
    )
vTopTableAll <-
  subset(
    vTopTableAll,
    select = -c(
      score,
      phase,
      transcript_id,
      transcript_type,
      transcript_name,
      transcript_support_level,
      exon_number,
      exon_id,
      ont,
      protein_id,
      ccdsid
    )
  )
    saveRDS(vTopTableAll, file = 
            paste0("../../rObjects/", directory, 
                   i,"_LBD_gene_table.rds"))
  path <- paste0("../../results/", tool, "/DEGs/",
  i, "_gene_DEGs_FDRq1.00.txt", sep = "") 
  write.table(
    vTopTableAll,
    path,
    sep = "\t",
    row.names = FALSE,
    quote = FALSE
  )
  # p < 0.05, log2fc > 0
  vTopTable1 <-
    topTable( 
      veBayesFit,  
      coef = coef,  
      n = Inf, 
      p.value = 0.05,
      lfc = 0
    )
  path <- paste0("../../results/", tool, "/DEGs/",
  i, "_gene_DEGs_FDRq0.05.txt", sep = "") 
  write.table(
    vTopTable1,
    path,
    sep = "\t",
    row.names = FALSE,
    quote = FALSE
  )
  # increment 
  coef <- coef + 1
}
remove(coef)
```

Read table with all genes (FDRq = 1).
```{r}
coef <- 1

for (i in allComparisons) {
  filepath <- paste0(
    "../../results/",
    tool,
    "/DEGs/",
    i,
    "_gene_DEGs_FDRq1.00.txt"
  )
  assign(i,
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ))
  # increment
  coef <- coef + 1
}
```
# assign colors
Assign colors  values based on FDRq cutoff of 0.05.
```{r assign_colors}
color_values <- vector()
max <- nrow(LBDvsControl)

for(i in 1:max){
  if (LBDvsControl$adj.P.Val[i] < 0.05){
    if (LBDvsControl$logFC[i] > 0){
      color_values <- c(color_values, 1) # 1 when logFC > 0 and FDRq < 0.05
    }
    else if (LBDvsControl$logFC[i] < 0){
      color_values <- c(color_values, 2) # 2 when logFC < 0 and FDRq < 0.05
    }
  }
  else{
    color_values <- c(color_values, 3) # 3 when FDRq >= 0.05
  }
}

LBDvsControl$color_p0.05 <- factor(color_values)
```

# subset genes to label
Subset the top 10 up and down-regulated genes
```{r}
up <- LBDvsControl[LBDvsControl$color_p0.05 == 1,]
up10 <- up[1:10,]

down <- LBDvsControl[LBDvsControl$color_p0.05 == 2,]
down <- subset(down, down$logFC < -1.5)
down10 <- down[1:7,]
```

# volcano plot
```{r volcano}
hadjpval <- (-log10(max(
  LBDvsControl$P.Value[LBDvsControl$adj.P.Val < 0.05], 
  na.rm=TRUE)))

p_vol <-
  ggplot(data = LBDvsControl, 
         aes(x = logFC,  # x-axis is logFC
             y = -log10(P.Value),  # y-axis will be -log10 of P.Value
             color = color_p0.05)) +  # color is based on factored color column
  geom_point(alpha = 1.5, size = 1.7) +  # create scatterplot, alpha makes points transparent
  theme_bw() +  # set color theme
  theme(legend.position = "none") +  # no legend
  scale_color_manual(values = c("red", "blue","grey")) +  # set factor colors
  labs(
    title = "", # no main title
    x = expression(log[2](FC)), # x-axis title
    y = expression(-log[10] ~ "(" ~ italic("p") ~ "-value)") # y-axis title
  ) +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  geom_hline(yintercept = hadjpval,  #  horizontal line
                     colour = "#000000",
                     linetype = "dashed") +
  ggtitle("LBD vs Control\nFDRq < 0.05") +
  theme(plot.title = element_text(size = 10)) +
  geom_text_repel(data = up10,
                  aes(x = logFC, y= -log10(P.Value), label = gene_name), 
                  color = "maroon", 
                  fontface="italic",
                  size = 4, 
                  max.overlaps = getOption("ggrepel.max.overlaps", default = 30)
                  ) +
  geom_text_repel(data = down10,
                  aes(x = logFC, y= -log10(P.Value), label = gene_name), 
                  color = "navyblue", 
                  fontface="italic",
                  size = 4, 
                  max.overlaps = getOption("ggrepel.max.overlaps", default = 15)
                  )  #+
 # scale_y_continuous(breaks = seq(0,8,by=1), limits = c(0,8)) +
 # scale_x_continuous(breaks = seq(-3,3,by=1), limits = c(-3,3))
p_vol

# save
#path <- paste0("../../results/", tool, "/volcano/LBD_gene_volcano_FDRq0.05")
#saveToPDF(paste0(path, ".pdf"), width = 5.2, height = 5.2)
```

# blood & white matter marker expression among groups 
```{r}
# get the gene id that corresponds to the gene name 
genes <- dge.filtered.norm$genes

bloodmarker_gene_info <- subset(genes, 
                                gene_name == "HBB" |  
                                gene_name == "HBA1" |
                                gene_name == "HBA2" |
                                gene_name == "HBD" |
                                gene_name == "HBQ1" |
                                gene_name == "MOG" |
                                gene_name == "MBP" |
                                gene_name == "RBFOX3")
bloodmarker_gene_id <- rownames(bloodmarker_gene_info)
# subset counts 
HBB_counts <- subset(voomCountsMatrix, rownames(voomCountsMatrix) %in% bloodmarker_gene_id)
HBB_melt <- reshape2::melt(HBB_counts)
names(HBB_melt)[names(HBB_melt) == "Var2"] <- "NPID"
names(HBB_melt)[names(HBB_melt) == "Var1"] <- "gene_id"
gene_id_gene_name <- data.frame(rownames(bloodmarker_gene_info), bloodmarker_gene_info$gene_name)
names(gene_id_gene_name)[names(gene_id_gene_name) == "rownames.bloodmarker_gene_info."] <- "gene_id"
names(gene_id_gene_name)[names(gene_id_gene_name) == "bloodmarker_gene_info.gene_name"] <- "gene_name"
df <- merge(gene_id_gene_name, HBB_melt, by = "gene_id")
df <- merge(df, info, by = "NPID")
```

```{r}
column_variables <- c("HBB", "HBA1", "HBA2", "HBD", "HBQ1", "MOG", "MBP", "RBFOX3")
my_comparisons <-
  list(
    c("CONTROL", "AD"),
    c("CONTROL", "PA"),
    c("CONTROL", "LBD"),
    c("AD", "PA"),
    c("AD", "LBD"),
    c("PA", "LBD")
  )
violin_plot_fun <- function(j) {
  df <- subset(df, gene_name == j)
  ggplot(df, aes(TYPE, value, fill = TYPE)) +
    geom_violin() +
    geom_boxplot(width = 0.1) + 
    geom_jitter(shape=16, position=position_jitter(0.2)) +
    theme_bw() + 
    ggtitle(j) +
    xlab("Disease group") +
    ylab(j) +
    stat_compare_means(comparisons = my_comparisons) +
    scale_fill_manual(values=TypeColors) +
    theme(legend.position = "none")
}
violin_plots <-
  Map(violin_plot_fun, j = column_variables)
violin_plots

plotnames = imap(violin_plots, ~paste0("../../results/", tool, "/blood_markers/Type.", .y, ".pdf")) %>%
  flatten()
plotnames
walk2(plotnames, violin_plots, ~ggsave(filename = .x, plot = .y, 
                                    height = 4, width = 5.5))


# blood marker expression between VaD classification 
column_variables <- c("HBB", "HBA1", "HBA2", "HBD", "HBQ1")
my_comparisons <- list(c("0", "1"))
df$VaD <- as.factor(df$VaD)
violin_plot_fun <- function(j) {
  df <- subset(df, gene_name == j)
  ggplot(df, aes(VaD, value, fill = VaD)) +
    geom_violin() +
    geom_boxplot(width = 0.1) + 
    geom_jitter(shape=16, position=position_jitter(0.2)) +
    theme_bw() + 
    ggtitle(j) +
    xlab("Disease group") +
    ylab(j) +
    stat_compare_means(comparisons = my_comparisons) +
    theme(legend.position = "none")
}
violin_plots <-
  Map(violin_plot_fun, j = column_variables)
violin_plots

plotnames = imap(violin_plots, ~paste0("../../results/", tool, "/blood_markers/VaD.", .y, ".pdf")) %>%
  flatten()
plotnames
walk2(plotnames, violin_plots, ~ggsave(filename = .x, plot = .y, 
                                    height = 4, width = 5.5))
```
