==> scripts/R/04c_create_SCC_counts_object.Rmd <==
---
title: "Create rObject of RNAseq counts data"
author: "Kimberly Olney, Ph.D."
date: "09/30/2022"
output:
  pdf_document: default
  html_document:
    df_print: paged
params:
  args: myarg
---
Samples have already been sex checked and realigned to the SCC reference. 
# Setup
```{r setup}
knitr::opts_knit$set(root.dir = ".")
```
# User defined variables
```{r set_variables}
source(here::here("scripts/R", "file_paths_and_colours.R"))
source(here::here("scripts/R", "gtf_path.R"))
condition <- c("") # change condition to rename files for comparing between different approaches. 
tool = c("star")
typeOfCount <-  c(".bamReadsPerGene.out.tab")
```
# Read data
### Metadata
```{r read_data}
# example of file naming:
# NA12-254_FCHCLTJDMXY_L1_STAR_XY.bamReadsPerGene.out.tab
# path to counts files
count_files <-
  file.path(paste0(
    "../../starAligned_SCC/",
    metadata$NPID,
    "_STAR_",
    metadata$sex_chr,
    typeOfCount
  ))
# add sample name to counts files
names(count_files) <- paste0(metadata$NPID)
# add sample count path information to metadata
metadata$path <- count_files
# make sure there is only unique paths
```
### Create a counts matrix
```{r counts_matrix}
counts <- data.frame(fread(count_files[1]))[c(1, 4)]
# Loop and read the remaining counts files
for (i in 2:length(count_files)) {
  counts <- cbind(counts, data.frame(fread(count_files[i]))[4])
}
```
# Star mapping metrics
```{r mapping_metrics}
star_metrics <- data.frame(counts[counts$V1 %like% "N_", ])
star_metrics$V1 <- as.vector(star_metrics$V1)
melt_star_metrics <- reshape::melt(star_metrics, id = c("V1"))
ggplot(melt_star_metrics, aes(x = value, fill = V1)) + 
  geom_histogram(position = "identity", bins = 100) + 
  facet_grid(V1 ~ .)  + theme_bw()

path <- paste0("../../results/", tool, "/mapping_metrics/star_mapping_metrics")
saveToPDF(paste0(path, ".pdf"), width = 7, height = 5)
```
clean up
```{r}
# clean up
remove(melt_star_metrics, 
       params, 
       star_metrics)
```
# Format counts table
```{r counts_table}
# remove star metric information
counts <- counts[!grepl("N_", counts$V1),]
# set gene_id (i.e. V1) as the row name
row.names(counts) <- counts$V1
counts$V1 <- NULL
# set column names to NPID sample ID
colnames(counts) <- metadata$NPID
```
# Check inputs
```{r check_inputs}
all.equal(rownames(counts), genes.gtf$gene_id)
all.equal(colnames(counts), (metadata$NPID))
```
# Create DGE object
```{r DGE_object}
# create object
metadata$group <- metadata$TYPE
dge <- DGEList(counts = counts,
               samples = metadata,
               genes = genes.gtf)

table(dge$samples$TYPE)
table(dge$samples$ATS_names)
table(dge$samples$group)
```

# Save R object
```{r rOjects}
saveRDS(dge, file = paste0("../../rObjects/", condition, "dge.raw.rds"))
dge <- readRDS(paste0("../../rObjects/", condition, "dge.raw.rds"))
```
# MDS
convert counts to cpm and lcpm
set colors and get data
```{r cpm, warning=FALSE}
lcpm <- edgeR::cpm(dge$counts, log = TRUE)
cpm <- edgeR::cpm(dge$counts, log = FALSE)

# sex colors 
dge$samples$sex_inferred <- as.factor(dge$samples$sex_inferred)
sex_colors <- c(SexColors)[dge$samples$sex_inferred]

par(bg = 'white')
plotMDS(
  lcpm,
  top = 100, 
  labels = dge$samples$sex_inferred,
  cex = .8, 
  dim.plot = c(1,2), 
  plot = TRUE, 
  col = sex_colors,
  gene.selection = "common"
)
title(expression('Top 100 Genes (Log'[2]~'CPM)'))

path <- paste0("../../results/", tool, "/MDS/", 
               condition,"raw_counts_label_sex_dim1&2")
saveToPDF(paste0(path, ".pdf"), width = 5.2, height = 5.2)

plotMDS(
  lcpm,
  top = 100, 
  labels = dge$samples$sex_inferred,
  cex = .8, 
  dim.plot = c(2,3), 
  plot = TRUE, 
  col = sex_colors,
  gene.selection = "common"
)
title(expression('Top 100 Genes (Log'[2]~'CPM)'))
path <- paste0("../../results/", tool, "/MDS/", 
               condition,"raw_counts_label_sex_dim2&3")
saveToPDF(paste0(path, ".pdf"), width = 5.2, height = 5.2)
```

# Filtering and RIN check
The filterByExpr() function in the edgeR package determines which genes have a 
great enough count value to keep.  
We will filter by group.  
This means at least 6 samples (6 is the smallest group sample size) must express a minimum count of 
1 (in cpm, default value is 10).
```{r filter_counts}
# first filter by expression
dim(dge)
keep.expr <-
  filterByExpr(
    edgeR::cpm(dge$counts),
    dge$samples$TYPE, # by ATS groups
    min.count = 1, # min count of 1 CPM 
    min.prop = 0.3 # at least 30% of the smallest group 
  )
dge.filtered <- dge[keep.expr, , keep.lib.sizes = FALSE]
dim(dge.filtered)
table(dge.filtered$genes$gene_type)
```
# Remove mitochondrial genes & keep only protein coding genes
```{r MT_genes, eval = TRUE}
dim(dge.filtered)
removeMT <- dge.filtered$genes$seqnames != "chrM"  # true when NOT MT
dge.filtered <- dge.filtered[removeMT,,keep.lib.sizes = FALSE]

keepProtein <- dge.filtered$genes$gene_type == "protein_coding"
dge.filtered <- dge.filtered[keepProtein,,keep.lib.sizes = FALSE]
dim(dge.filtered)
```

```{r eval = FALSE}
cpm <- edgeR::cpm(dge.filtered$counts, log = FALSE)
write.table(cpm, "/research/labs/neurology/fryer/m239830/LBD_CWOW/results/star/counts/cpm.tsv", quote = FALSE, sep = "\t")
dim(cpm)

lcpm <- edgeR::cpm(dge.filtered$counts, log = TRUE)
write.table(lcpm, "/research/labs/neurology/fryer/m239830/LBD_CWOW/results/star/counts/lcpm.tsv", quote = FALSE, sep = "\t")
write.table(metadata, "/research/labs/neurology/fryer/m239830/LBD_CWOW/results/star/counts/metadata.tsv", quote = FALSE, sep = "\t")
```
# Library sizes
```{r library}
# before filtering
L <- mean(dge$samples$lib.size) * 1e-6
M <- median(dge$samples$lib.size) * 1e-6
c(L, M)

# after filtering 
L <- mean(dge.filtered$samples$lib.size) * 1e-6
M <- median(dge.filtered$samples$lib.size) * 1e-6
c(L, M)
```

# TMM
For estimating relative RNA production levels from RNA-seq data. 
The TMM method estimates scale factors between samples that can be incorporated 
into currently used statistical methods for DE analysis.
```{r TMM}
# Now, method of trimmed mean of M-values (TMM)
# Calculate scaling factors to convert raw library sizes into effective library sizes.
dge.filtered.norm <- calcNormFactors(dge.filtered, method = "TMM")

# norm factor summary
summary(dge.filtered.norm$samples$norm.factors)
normfactors <- (dge.filtered.norm$samples$norm.factors)
metadata$normfactors <- normfactors

# examine normalization factors 
plot(metadata$TYPE, metadata$normfactors)
plot(metadata$ATS_names, metadata$normfactors)

ggplot(metadata, aes(TYPE, normfactors, fill = TYPE)) +
  geom_violin(alpha = 0.5) +
  geom_boxplot(width = 0.1, outlier.shape = NA, color = "black") +
  geom_jitter(shape = 16, position = position_jitter(0.2), alpha = 0.2) +
  theme_bw() +
  ggtitle("Normalization factors") +
  xlab("TYPE") +
  ylab("TMM normalization factor") +
  scale_fill_manual(values = TypeColors) +
  theme(legend.position = "none")

path <- paste0("../../results/", tool, "/library/TYPE.TMM")
saveToPDF(paste0(path, ".pdf"), width = 5.5, height = 4)
```

```{r plot_RIN}
log2cpm.norm <- edgeR::cpm(dge.filtered.norm, log = T)
nsamples <- ncol(dge.filtered.norm)
boxplot(log2cpm.norm, 
        main="Filtered normalized lcpm data", 
        xlab="RIN", 
        ylab=expression('Counts per gene (Log'[2]~'CPM)'),
        axes=FALSE)
axis(2)
axis(1,at=1:nsamples,labels=(dge.filtered.norm$samples$RIN),las=2,cex.axis=0.8)

path <- paste0("../../results/", tool, "/library/", condition,
               "boxplot_RIN")
saveToPDF(paste0(path, ".pdf"), width = 35, height = 6)
```

# RIN check and library size
```{r RIN_check, warning=FALSE}
# check if there is correlation between RIN and library size
box <- dge.filtered.norm$samples
box <- box[!is.na(box$RIN),]
cor(box$RIN, box$lib.size, method = c("pearson", "kendall", "spearman"))
cor.test(box$RIN, box$lib.size, method=c("pearson", "kendall", "spearman"))

# is the data normally distributed 
ggqqplot(box$lib.size, ylab = "library size")
ggqqplot(box$RIN, ylab = "RIN")
res <- cor.test(box$lib.size, box$RIN, 
                method = "spearman")
res
ggscatter(box, x = "RIN", y = "lib.size",
          add = "reg.line", conf.int = TRUE, color = "black",
          cor.coef = TRUE, cor.method = "spearman",
          ylab = "library size", xlab = "RIN value") +
  ggtitle("RIN and library size")

path <- paste0("../../results/", tool, "/library/", condition, 
               "corr_RIN_library_size")
saveToPDF(paste0(path, ".pdf"), width = 6, height = 4)


ggscatter(box, x = "RIN", y = "lib.size", fill = "TYPE",
          add = "reg.line", conf.int = TRUE, color = "black",
          size = .5, cor.coef = TRUE, cor.method = "spearman",
          ylab = "library size", xlab = "RIN value") +
  ggtitle("RIN and library size") + facet_grid(. ~ TYPE) +
  scale_fill_manual(values=TypeColors) +
  theme(legend.position = "none")
path <- paste0("../../results/", tool, "/library/", condition, 
               "TYPE.corr_RIN_library_size")
saveToPDF(paste0(path, ".pdf"), width = 8, height = 4)
```

# MDS - normalized counts
multidimensional scaling
```{r MDS_norm}
par(bg = 'white')
plotMDS(
  log2cpm.norm,
  top = 100, 
  labels = dge.filtered.norm$samples$sex_inferred,
  cex = .8, 
  dim.plot = c(1,2), 
  plot = TRUE, 
  col = sex_colors,
  gene.selection = "common"
)
title(expression('Top 100 Genes (Log'[2]~'CPM)'))

path <- paste0("../../results/", tool, "/MDS/", 
               condition, "filtered_counts_label_sex_dim1&2")
saveToPDF(paste0(path, ".pdf"), width = 5.2, height = 5.2)
```
# Density plot
Density plots of log-intensity distribution of each library can be superposed on a single graph for a better comparison between libraries and for identification of libraries with weird distribution. 
```{r density_plots}
# Normalize data for library size and expression intensity
log2cpm.tech <- edgeR::cpm(dge, log = TRUE)
log2cpm.filtered <- edgeR::cpm(dge.filtered, log = TRUE)
log2cpm.norm <- edgeR::cpm(dge.filtered.norm, log = TRUE)

# set graphical parameter
par(mfrow = c(1,3))
# set colors
colors <- TypeColors[dge$samples$TYPE]
nsamples <- ncol(dge)

# First, plot the first column of the log2cpm.tech density
plot(density(log2cpm.tech[,1]), col = colors[1], lwd = 2, ylim = c(0,0.5), 
     las = 2, main = "A. Raw", xlab = expression('Log'[2]~CPM))
# For each sample plot the lcpm density
for (i in 1:nsamples){
  den <- density(log2cpm.tech[,i]) #subset each column
  lines(den$x, den$y, col = colors[i], lwd = 2) 
}

# Second, plot log2cpm.filtered
plot(density(log2cpm.filtered[,1]), col = colors[1], lwd = 2, ylim = c(0,0.5), 
     las = 2, main = "B. Filtered", xlab = expression('Log'[2]~CPM))
for (i in 2:nsamples) {
  den <- density(log2cpm.filtered[,i])
  lines(den$x, den$y, col = colors[i], lwd = 2)
}

# Third, plot log2cpm.norm
plot(density(log2cpm.norm[,1]), col = colors[1], lwd = 2, ylim = c(0,0.5), 
     las = 2, main = "C. TMM", xlab = expression('Log'[2]~CPM))
for (i in 2:nsamples) {
  den <- density(log2cpm.norm[,i])
  lines(den$x, den$y, col = colors[i], lwd = 2)
}

# save
path <- paste0("../../results/", tool, "/library/TYPE.gene_density")
saveToPDF(paste0(path, ".pdf"), width = 10, height = 5)

# set graphical parameter
par(mfrow = c(1,3))
# set colors
colors <- colorbindColors[dge$samples$ATS_names]
nsamples <- ncol(dge)
# First, plot the first column of the log2cpm.tech density
plot(density(log2cpm.tech[,1]), col = colors[1], lwd = 2, ylim = c(0,0.5), 
     las = 2, main = "A. Raw", xlab = expression('Log'[2]~CPM))
# For each sample plot the lcpm density
for (i in 2:nsamples){
  den <- density(log2cpm.tech[,i]) #subset each column
  lines(den$x, den$y, col = colors[i], lwd = 2) 
}
# Second, plot log2cpm.filtered
plot(density(log2cpm.filtered[,1]), col = colors[1], lwd = 2, ylim = c(0,0.5), 
     las = 2, main = "B. Filtered", xlab = expression('Log'[2]~CPM))
for (i in 2:nsamples) {
  den <- density(log2cpm.filtered[,i])
  lines(den$x, den$y, col = colors[i], lwd = 2)
}
# Third, plot log2cpm.norm
plot(density(log2cpm.norm[,1]), col = colors[1], lwd = 2, ylim = c(0,0.5), 
     las = 2, main = "C. TMM", xlab = expression('Log'[2]~CPM))
for (i in 2:nsamples) {
  den <- density(log2cpm.norm[,i])
  lines(den$x, den$y, col = colors[i], lwd = 2)
}
# save
path <- paste0("../../results/", tool, "/library/ATS.gene_density")
saveToPDF(paste0(path, ".pdf"), width = 7, height = 5)
```
clean up 
```{r}
remove(box,
       res)
```

# GOI gene of interest histogram
```{r GOI}
lcpm <- edgeR::cpm(dge.filtered.norm$counts, log = TRUE)
cpm <- edgeR::cpm(dge.filtered.norm$counts, log = FALSE)

# using CPM data
cpm <- as.data.frame(cpm)
cpm$gene_id  <- rownames(cpm)
cpm_genes <- merge(cpm, dge.filtered.norm$genes, by = "gene_id")
goi <- subset(cpm_genes, gene_name == "TTR")
goi <- as.vector(goi[1:620]) # get only sample counts data. Columns 2-620 (i.e. 619 samples)
goi_melt <- melt(goi) # transform 
summary(goi_melt$value) 
colnames(goi_melt)[2] <- "NPID"
goi_melt_sample_info <-
  merge(goi_melt, dge.filtered.norm$samples, by = "NPID")
df <-
  with(goi_melt_sample_info, goi_melt_sample_info[order(TYPE, NPID), ])
df$NPID <- factor(df$NPID, levels = df$NPID)
# column plot
ggplot(df, aes(x = NPID, y = value, fill = TYPE)) +
  geom_col() +
  labs(title = "TTR", x = NULL, y = NULL) +
  xlab("TYPE") + ylab("CPM") +
  theme_bw() +
  scale_fill_manual(values = TypeColors)
path <- paste0("../../results/", tool, "/goi/TTR.CPM.bar")
saveToPDF(paste0(path, ".pdf"), width = 8, height = 5)

# using log2CPM data
lcpm <- as.data.frame(lcpm)
lcpm$gene_id  <- rownames(lcpm)
lcpm_genes <- merge(lcpm, dge.filtered.norm$genes, by = "gene_id")
goi <- subset(lcpm_genes, gene_name == "TTR")
goi <- as.vector(goi[1:620])
goi_melt <- melt(goi)
summary(goi_melt$value)
colnames(goi_melt)[2] <- "NPID"
goi_melt_sample_info <-
  merge(goi_melt, dge.filtered.norm$samples, by = "NPID")
df <-
  with(goi_melt_sample_info, goi_melt_sample_info[order(TYPE, NPID), ])
df$NPID <- factor(df$NPID, levels = df$NPID)
ggplot(df, aes(x = NPID, y = value, fill = TYPE)) +
  geom_col() +
  labs(title = "TTR", x = NULL, y = NULL) +
  xlab("TYPE") + ylab("log2(CPM)") +
  theme_bw() +
  scale_fill_manual(values = TypeColors)
path <- paste0("../../results/", tool, "/goi/TTR.log2CPM.bar")
saveToPDF(paste0(path, ".pdf"), width = 8, height = 5)
```
clean up 
```{r}
remove(
       dge, 
       dge.filtered,
       lcpm,
       log2cpm.filtered,
       log2cpm.norm,
       log2cpm.tech)
```

# Save R object
```{r save_dge_final}
saveRDS(dge.filtered.norm, file = paste0("../../rObjects/", condition, "dge.filtered.norm.rds"))
```

```{r}
sessionInfo()
```
