==> scripts/R/07_gprofiler_TYPE_SEX.Rmd <==
---
title: "gprofiler2"
author: "Kimberly Olney"
date: "05/18/2023"
output: html_document
---
# gprofiler for gene enrichment analysis
A web version of gprofiler is available here: https://biit.cs.ut.ee/gprofiler/gost 
A vignette is here: https://cran.r-project.org/web/packages/gprofiler2/vignettes/gprofiler2.html#gene-identifier-conversion-with-gconvert 
```{r setup, include=FALSE, echo=FALSE}
knitr::opts_knit$set(root.dir = ".")
```

# User defined variables
```{r set_variables, warning=FALSE, echo=FALSE}
source(here::here("/research/labs/neurology/fryer/m239830/LBD_CWOW/scripts/R", "file_paths_and_colours.R"))
condition <- c("TYPE.SEX") #  TYPE
tool <- c("star")
```

# Read in DEGs
The DEGs are already ordered by smallest to largest adjusted p-value.
```{r loop, echo=FALSE}
allComparisons <- c("male_XY_LBDvsControl", "male_XY_ADvsControl", "male_XY_LBDvsAD") 
# first read in the files
for (i in allComparisons) {
  filepath <- paste0("../../results/",tool,
    "/DEGs/", condition, "_", i,
    "_gene_DEGs_FDRq0.05_logFC_0.25.txt")
  df <- assign(paste0(i),
         tryCatch(
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ), error=function(e) NULL))
  up <- subset(df$gene_name, df$logFC > 0, sort = FALSE)
  assign(paste0(i, "_up"), up)
  down <- subset(df$gene_name, df$logFC < 0, sort = FALSE)
  assign(paste0(i, "_down"), down)
  
  # gprofiler
  gp_up <- gost(up, ordered_query = TRUE, organism = "hsapiens")
  gp_down <- gost(down, ordered_query = TRUE, organism = "hsapiens")

  multi_gp = gost(list("down-regulated" = down, 
                     "up-regulated" = up), ordered_query = TRUE, 
                     organism = "hsapiens")
  # rearrange so that up-regulated is on top 
  neworder <- c("up-regulated","down-regulated")
  multi_gp$result <- arrange(transform(multi_gp$result,
             query=factor(query,levels=neworder)),query)
  print(paste("i =", i))
  # interactive plot 
  gostplot(multi_gp, interactive = TRUE)
  p_up_and_down = gostplot(multi_gp, interactive = FALSE)
  p <- publish_gostplot(p_up_and_down)
  path <- paste0("../../results/", tool, "/gprofiler/", i, "_up_and_down_gprofiler")
  pdf(paste0(path, ".pdf"), height = 6, width = 8)
  print(p)
  dev.off()
  
  # needed to convert to enrichResult object
  up_names = gconvert(up, organism = "hsapiens") 
  down_names = gconvert(down,organism = "hsapiens")
  # enrichment analysis using gene names
  multi_gp = gost(list("up-regulated" = up_names$name, "down-regulated" = down_names$name), 
                  multi_query = FALSE, evcodes = TRUE, ordered_query = TRUE, organism = "hsapiens")
  write.table(up_names, paste0("../../results/", tool,"/gprofiler/", 
                               i, "_up_gene_descriptions.txt"), 
              sep = "\t", quote = FALSE, row.names = FALSE)
  write.table(down_names, paste0("../../results/", tool,"/gprofiler/", 
                               i, "_down_gene_descriptions.txt"), 
              sep = "\t", quote = FALSE, row.names = FALSE)
  # reorder so that up-regulated is on top
  neworder <- c("up-regulated","down-regulated")
  multi_gp$result <- arrange(transform(multi_gp$result,
             query=factor(query,levels=neworder)),query)

  # modify the g:Profiler data frame
  gp_mod = multi_gp$result[,c("query", "source", "term_id", "term_name", "p_value", 
                            "query_size", "intersection_size","term_size", 
                            "effective_domain_size","intersection")] 
  # get the gene name size for each GO 
  gp_mod$GeneRatio = paste0(gp_mod$intersection_size, "/", gp_mod$query_size)
  gp_mod$BgRatio = paste0(gp_mod$term_size, "/", gp_mod$effective_domain_size) 
  names(gp_mod) = c("Cluster", "Category", "ID", "Description", "p.adjust", 
                    "query_size", "Count", "term_size", "effective_domain_size",
                    "gene_name", "GeneRatio", "BgRatio") 
  gp_mod$gene_name = gsub(",", "/", gp_mod$gene_name)
  row.names(gp_mod) = make.names(gp_mod$ID, unique = TRUE)
  # define as compareClusterResult object
  gp_mod_cluster = new("compareClusterResult", compareClusterResult = gp_mod)
  # define as enrichResult object
  gp_mod_enrich = new("enrichResult", result = gp_mod)
  # output table 
  write.table(gp_mod_enrich@result, paste0("../../results/", tool,
                                           "/gprofiler/", i, "_up_and_down_gp_mod_enrich.txt"), 
              sep = "\t", quote = FALSE, row.names = FALSE)
}
```
# Make excel table
```{r excel, echo=FALSE}
# first read in the files
for (i in allComparisons) {
  filepath <- paste0("../../results/",tool,
    "/gprofiler/", i,
    "_up_and_down_gp_mod_enrich.txt"
  )
  assign(paste0(i),
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ))
}
allComparisons
list_of_datasets <- list(
"male_XY_ADvsControl" = male_XY_ADvsControl,     
"male_XY_LBDvsControl" = male_XY_LBDvsControl,
"male_XY_LBDvsAD "= male_XY_LBDvsAD)
write.xlsx(list_of_datasets, file = paste0(
    "../../results/",
    tool,
    "/gprofiler/", condition,"_male_XY_up_and_down_gp_mod_enrich.xlsx"))
```
# Excel gene name descriptions
```{r}
# first read in the files
for (i in allComparisons) {
  filepath <- paste0("../../results/",tool,
    "/gprofiler/", i,
    "_up_gene_descriptions.txt"
  )
  assign(paste0(i),
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ))
}
list_of_datasets <- list(
"male_XY_ADvsControl" = male_XY_ADvsControl,     
"male_XY_LBDvsControl" = male_XY_LBDvsControl,
"male_XY_LBDvsAD "= male_XY_LBDvsAD)
write.xlsx(list_of_datasets, file = paste0(
    "../../results/",
    tool,
    "/gprofiler/", condition,"_male_XY_up_gene_descriptions.txt.xlsx"))

# first read in the files
for (i in allComparisons) {
  filepath <- paste0("../../results/",tool,
    "/gprofiler/", i,
    "_down_gene_descriptions.txt"
  )
  assign(paste0(i),
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ))
}
write.xlsx(list_of_datasets, file = paste0(
    "../../results/",
    tool,
    "/gprofiler/", condition,"_male_XY_down_gene_descriptions.txt.xlsx"))
```
# GO Plot
```{r plot, eval=FALSE, echo=FALSE}
up <- subset(ADvsControl, Cluster == "up-regulated")
up <- up[1:15, ]
p_enrich <- ggplot(data=up, aes(x=Count, y=Description)) +
   geom_bar(stat="identity", aes(fill = p.adjust), width = .65) +
  theme(axis.title.x = element_text(size = 12),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 12),
        axis.text.y = element_text(size = 10)) +
    theme_bw() +  # set color theme
  #  scale_y_discrete(expand=c(0,0)) +
  scale_fill_gradientn(colours = c("darkmagenta", "mediumorchid3", "grey"), values = rescale(c(0,.01, 0.05)), guide = "legend", limits=c(0,0.05)) + labs(x = "Gene count", y = "")
p_enrich
```
