==> scripts/R/07_upsets_and_correlations_sex_interaction.Rmd <==
---
title: "Upset and correlations among pairwise DEGs"
author: "Kimberly Olney"
date: "02/14/2023"
output: html_document
---

# Setup
```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = ".")
```
# Library packages
```{r packages}
library(UpSetR)
library(ComplexUpset)
library(ggplot2movies)
library(ggvenn)
library(forcats)
```
# Function to create tables
```{r}
fromList <- function (input) {
  # Same as original fromList()...
  elements <- unique(unlist(input))
  data <- unlist(lapply(input, function(x) {
      x <- as.vector(match(elements, x))
      }))
  data[is.na(data)] <- as.integer(0)
  data[data != 0] <- as.integer(1)
  data <- data.frame(matrix(data, ncol = length(input), byrow = F))
  data <- data[which(rowSums(data) != 0), ]
  names(data) <- names(input)
  # ... Except now it conserves your original value names!
  row.names(data) <- elements
  return(data)
}

makePaddedDataFrame <- function(l, ...) {
  maxlen <- max(sapply(l, length))
  data.frame(lapply(l, na.pad, len = maxlen), ...)
}
```
# User defined variables
```{r set_variables}
source(here::here("scripts/R", "file_paths_and_colours.R"))
condition <- c("TYPE.SEX")
tool <- "star"
# save contrast names
allComparisons <- c("LBDvsControl", 
                    "LBDvsAD", 
                    "LBDvsPA", 
                    "ADvsControl", 
                    "PAvsControl", 
                    "ADvsPA")
allComparisons # check
```
# Read in DEGs
Read table with all genes (FDRq = 1).
```{r read_DEG_table}
# sex as a factor
coef <- 1
condition <- c("TYPE")
for (i in allComparisons) {
  filepath <- paste0(
    "../../results/",
    tool,
    "/DEGs/", condition, "_", 
    i,
    "_gene_DEGs_FDRq1.00.txt"
  )
  assign(paste0(i),
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ))
  # increment
  coef <- coef + 1
}

# XX females only 
coef <- 1
sex_condition <- c("female_XX")
condition <- c("TYPE.SEX")
for (i in allComparisons) {
  filepath <- paste0(
    "../../results/",
    tool,
    "/DEGs/", condition, "_",  sex_condition, "_",
    i,
    "_gene_DEGs_FDRq1.00.txt"
  )
  assign(paste0(i,"_", sex_condition),
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ))
  # increment
  coef <- coef + 1
}

# XY males only 
coef <- 1
sex_condition <- c("male_XY")
for (i in allComparisons) {
  filepath <- paste0(
    "../../results/",
    tool,
    "/DEGs/",condition, "_", sex_condition, "_",
    i,
    "_gene_DEGs_FDRq1.00.txt"
  )
  assign(paste0(i,"_", sex_condition),
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ))
  # increment
  coef <- coef + 1
}

# difference of differences, sex interaction
coef <- 1
sex_condition <- c("sex_interaction")
for (i in allComparisons) {
  filepath <- paste0(
    "../../results/",
    tool,
    "/DEGs/",condition, "_",
    i,  "_", sex_condition, 
    "_gene_DEGs_FDRq1.00.txt"
  )
  assign(paste0(i,"_", sex_condition),
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ))
  # increment
  coef <- coef + 1
}

# sex differences within disease type
coef <- 1
allComparisons_sexDiff <- c("female_LBD_vs_male_LBD", "female_AD_vs_male_AD", "female_Control_vs_male_Control", 
                    "female_PA_vs_male_PA")
for (i in allComparisons_sexDiff) {
  filepath <- paste0(
    "../../results/",
    tool,
    "/DEGs/",condition, "_",
    i, 
    "_gene_DEGs_FDRq1.00.txt"
  )
  assign(paste0(i),
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ))
  # increment
  coef <- coef + 1
}
```

# DEGs by FDRq <= 0.05 & log2FC
```{r}
# sex as a factor
LBDvsControl_up <- subset(LBDvsControl, logFC > 0.25 & adj.P.Val <= 0.05)
LBDvsControl_down <- subset(LBDvsControl, logFC < -0.25 & adj.P.Val <= 0.05)
ADvsControl_up <- subset(ADvsControl, logFC > 0.25 & adj.P.Val <= 0.05)
ADvsControl_down <- subset(ADvsControl, logFC < -0.25 & adj.P.Val <= 0.05)
PAvsControl_up <- subset(PAvsControl, logFC > 0.25 & adj.P.Val <= 0.05)
PAvsControl_down <- subset(PAvsControl, logFC < -0.25 & adj.P.Val <= 0.05)

# XX females
LBDvsControl_female_XX_up <- subset(LBDvsControl_female_XX, logFC > 0.25 & adj.P.Val <= 0.05)
LBDvsControl_female_XX_down <- subset(LBDvsControl_female_XX, logFC < -0.25 & adj.P.Val <= 0.05)
ADvsControl_female_XX_up <- subset(ADvsControl_female_XX, logFC > 0.25 & adj.P.Val <= 0.05)
ADvsControl_female_XX_down <- subset(ADvsControl_female_XX, logFC < -0.25 & adj.P.Val <= 0.05)
PAvsControl_female_XX_up <- subset(PAvsControl_female_XX, logFC > 0.25 & adj.P.Val <= 0.05)
PAvsControl_female_XX_down <- subset(PAvsControl_female_XX, logFC < -0.25 & adj.P.Val <= 0.05)

# XY males
LBDvsControl_male_XY_up <- subset(LBDvsControl_male_XY, logFC > 0.25 & adj.P.Val <= 0.05)
LBDvsControl_male_XY_down <- subset(LBDvsControl_male_XY, logFC < -0.25 & adj.P.Val <= 0.05)
ADvsControl_male_XY_up <- subset(ADvsControl_male_XY, logFC > 0.25 & adj.P.Val <= 0.05)
ADvsControl_male_XY_down <- subset(ADvsControl_male_XY, logFC < -0.25 & adj.P.Val <= 0.05)
PAvsControl_male_XY_up <- subset(PAvsControl_male_XY, logFC > 0.25 & adj.P.Val <= 0.05)
PAvsControl_male_XY_down <- subset(PAvsControl_male_XY, logFC < -0.25 & adj.P.Val <= 0.05)

# sex effect interaction 
LBDvsControl_interaction_up <- subset(LBDvsControl_sex_interaction, logFC > 0.25 & CI.L > 0 )
LBDvsControl_interaction_down <- subset(LBDvsControl_sex_interaction, logFC < -0.25 & CI.R < 0)
ADvsControl_interaction_up <- subset(ADvsControl_sex_interaction, logFC > 0.25 & CI.L > 0 )
ADvsControl_interaction_down <- subset(ADvsControl_sex_interaction, logFC < -0.25 & CI.R < 0)
PAvsControl_interaction_up <- subset(PAvsControl_sex_interaction, logFC > 0.25 & CI.L > 0 )
PAvsControl_interaction_down <- subset(PAvsControl_sex_interaction, logFC < -0.25 & CI.R < 0)
```

# Upset of DEGs
### LBD vs Control bt the sexes 
```{r}
list_input <- list("female down-regulated" = LBDvsControl_female_XX_down$gene_name,
                   "female up-regulated" = LBDvsControl_female_XX_up$gene_name,
                   "male down-regulated" = LBDvsControl_male_XY_down$gene_name,
                   "male up-regulated" = LBDvsControl_male_XY_up$gene_name)
data <- fromList(list_input)

upset_gene <- upset(data, set_sizes = FALSE,
      c('female down-regulated',
        'male down-regulated',
        'female up-regulated',
        'male up-regulated'),
      themes=upset_modify_themes(
        list('intersections_matrix'=theme(text=element_text(size=10),
              plot.margin = margin(0, 0, 0, 0, "cm")),
          'overall_sizes' = theme(axis.text.x = element_text(size =10), 
               plot.margin = margin(0, 0, 0, 0, "cm")))),
  queries=list(
    upset_query(set='female down-regulated', fill='blue'),
    upset_query(set='male down-regulated', fill='blue'),
    upset_query(set='female up-regulated', fill='red'),
    upset_query(set='male up-regulated', fill='red'),
    upset_query(intersect = c('female down-regulated','male down-regulated'), 
               fill='blue', only_components = 
                  c('intersections_matrix', 'Intersection size')),
    upset_query(intersect = c('female up-regulated','male up-regulated'), 
                fill='red', only_components = 
                  c('intersections_matrix', 'Intersection size'))
  ),
  intersections = list(c('female down-regulated', 'male down-regulated'), 
                       'female down-regulated', 
                       'male down-regulated',
                       c('female up-regulated', 'male up-regulated'), 
                       'female up-regulated', 
                       'male up-regulated',
                       c('female up-regulated', 'male down-regulated'), 
                       c('female down-regulated', 'male up-regulated')
  ),
  base_annotations=list(
    'Intersection size'=(
        intersection_size(
        size=2,
        text = list(size = 3),
        text_mapping = aes(),
        bar_number_threshold=3,  # show all numbers on top of bars
        width=0.3,   # reduce width of the bars
        mapping=aes(fill='bars_color')
      )
      + scale_fill_manual(values=c('grey'), guide='none')
      + scale_y_continuous(expand=expansion(mult=c(0, 0.5)))
      + theme(axis.text.y = element_text(size = 10),
              axis.title.y = element_text(size = 10),
          panel.grid.major=element_blank(),
          panel.grid.minor=element_blank(),
          plot.margin = margin(0, 0, 0, 0, "cm"), 
          axis.line=element_line(colour='black')))),
  matrix=intersection_matrix(
      geom=geom_point(
        shape='circle filled',
        size=2,
        stroke=0.45)) +
        theme(axis.text.y = element_text(size = 10),
              axis.title.y = element_text(size = 10), 
              plot.margin = margin(0, 0, 0, 0, "cm")),
    sort_sets=FALSE,
  sort_intersections=FALSE)
upset_gene

# save
path <- paste0("../../results/", tool, 
               "/upset/", condition, "_LBDvsControl_bt_sexes_FDRq0.05")
saveToPDF(paste0(path, ".pdf"), width = 5, height = 4)

# Binary table with colnames:
write.table(
  data,
  paste0("../../results/",
  tool,
  "/upset/", condition,
  "_LBDvsControl_bt_sexes_list_binary_results.txt"),
  sep = "\t",
  quote = FALSE
)
```

```{r}
female_down_regulated <-
  subset(
    data,
    `female down-regulated` == 1 &
      `female up-regulated` == 0  &
      `male down-regulated` == 0 &
      `male up-regulated` == 0
  )

female_up_regulated <-
  subset(
    data,
    `female down-regulated` == 0 &
      `female up-regulated` == 1  &
      `male down-regulated` == 0 &
      `male up-regulated` == 0
  )

male_down_regulated <-
  subset(
    data,
    `female down-regulated` == 0 &
      `female up-regulated` == 0  &
      `male down-regulated` == 1 &
      `male up-regulated` == 0
  )

male_up_regulated <-
  subset(
    data,
    `female down-regulated` == 0 &
      `female up-regulated` == 0  &
      `male down-regulated` == 0 &
      `male up-regulated` == 1
  )

na.pad <- function(x, len) {
  x[1:len]
}

df <-
  makePaddedDataFrame(
    list(
      "female down-regulated" = row.names(female_down_regulated),
      "female up-regulated" = row.names(female_up_regulated),
      "male down-regulated" = row.names(male_down_regulated),
      "male up-regulated" = row.names(male_up_regulated)
    )
  )
df[is.na(df)] <- ""
write.table(
  df,
  paste0("../../results/",
  tool,
  "/upset/", condition, 
  "_LBDvsControl_bt_sexes_list_gene_names.txt"
), row.names = FALSE, quote = FALSE, sep = "\t")
```

### LBD vs Control with sex as a factor 
```{r}
list_input <- list("sex factor down-regulated" = LBDvsControl_down$gene_name,
                   "sex factor up-regulated" = LBDvsControl_up$gene_name,
                   "female down-regulated" = LBDvsControl_female_XX_down$gene_name,
                   "female up-regulated" = LBDvsControl_female_XX_up$gene_name,
                   "male down-regulated" = LBDvsControl_male_XY_down$gene_name,
                   "male up-regulated" = LBDvsControl_male_XY_up$gene_name)
data <- fromList(list_input)

upset_gene <- upset(data, set_sizes = FALSE,
      c('sex factor down-regulated',
        'female down-regulated',
        'male down-regulated',
        'sex factor up-regulated',
        'female up-regulated',
        'male up-regulated'),
      themes=upset_modify_themes(
        list('intersections_matrix'=theme(text=element_text(size=10),
              plot.margin = margin(0, 0, 0, 0, "cm")),
          'overall_sizes' = theme(axis.text.x = element_text(size =10), 
              plot.margin = margin(0, 0, 0, 0, "cm")))),
  queries=list(
    upset_query(set='sex factor down-regulated', fill='blue'),
    upset_query(set='female down-regulated', fill='blue'),
    upset_query(set='male down-regulated', fill='blue'),
    upset_query(set='sex factor up-regulated', fill='red'),
    upset_query(set='female up-regulated', fill='red'),
    upset_query(set='male up-regulated', fill='red')
  ),
  intersections = list(c('sex factor down-regulated','female down-regulated', 'male down-regulated'), 
                       c('sex factor down-regulated', 'female down-regulated'), 
                       c('sex factor down-regulated', 'male down-regulated'), 
                       c('female down-regulated', 'male down-regulated'), 
                       'sex factor down-regulated',
                       'female down-regulated', 
                       'male down-regulated',
                       c('sex factor up-regulated','female up-regulated', 'male up-regulated'), 
                       c('sex factor up-regulated', 'female up-regulated'), 
                       c('sex factor up-regulated', 'male up-regulated'), 
                       c('female up-regulated', 'male up-regulated'), 
                       'sex factor up-regulated',
                       'female up-regulated', 
                       'male up-regulated',
                       c('female up-regulated', 'male down-regulated'), 
                       c('female down-regulated', 'male up-regulated')
  ),
  base_annotations=list(
    'Intersection size'=(
        intersection_size(
        size=2,
        text = list(size = 3),
        text_mapping = aes(),
        bar_number_threshold=3,  # show all numbers on top of bars
        width=0.3,   # reduce width of the bars
        mapping=aes(fill='bars_color')
      )
      + scale_fill_manual(values=c('grey'), guide='none')
      + scale_y_continuous(expand=expansion(mult=c(0, 0.5)))
      + theme(axis.text.y = element_text(size = 10),
              axis.title.y = element_text(size = 10),
          panel.grid.major=element_blank(),
          panel.grid.minor=element_blank(),
          plot.margin = margin(0, 0, 0, 0, "cm"), 
          axis.line=element_line(colour='black')))),
  matrix=intersection_matrix(
      geom=geom_point(
        shape='circle filled',
        size=2,
        stroke=0.45)) +
        theme(axis.text.y = element_text(size = 10),
              axis.title.y = element_text(size = 10), 
              plot.margin = margin(0, 0, 0, 0, "cm")),
    sort_sets=FALSE,
  sort_intersections=FALSE)
upset_gene

# save
path <- paste0("../../results/", tool, 
               "/upset/", condition, "_LBDvsControl_sex_factor_FDRq0.05")
saveToPDF(paste0(path, ".pdf"), width = 6, height = 4)

upset(data, intersect = c('sex factor down-regulated',
        'female down-regulated',
        'male down-regulated',
        'sex factor up-regulated',
        'female up-regulated',
        'male up-regulated'))

path <- paste0("../../results/", tool,
               "/upset/", condition, "_LBDvsControl_sex_factor_FDRq0.05_simple")
saveToPDF(paste0(path, ".pdf"), width = 9, height = 6)

# Binary table with colnames:
write.table(
  data,
  paste0("../../results/",
  tool,
  "/upset/", condition, 
  "_LBDvsControl_sex_factor_list_binary_results.txt"),
  sep = "\t",
  quote = FALSE
)
```

### LBD DEGs specific to each sex 
```{r}
# merge the male and female data frames 
# this will get the logFC for the female and male only analysis
LBD_FandM <-
  merge(LBDvsControl_female_XX,
        LBDvsControl_male_XY,
        by = "gene_id",
        all = T)
# remove unnecessary columns 
LBD_FandM <- LBD_FandM[c(7,1,2,6,8:15,22:29)]
colnames(LBD_FandM) = c("gene_name", "gene_id","chr", "gene_type", "female_logFC", "female_CI.L", "female_CI.R", "female_AveExpr", "female_t","female_P.val", "female_adj.P.Val", "female_B", "male_logFC", "male_CI.L", "male_CI.R", "male_AveExpr", "male_t","male_P.val", "male_adj.P.Val", "male_B")
# uniquely up or down in the females 
# not DE in the male only or sex as a factor
LBD_uniquely_up_females <-
  subset(data,
    `female up-regulated` == 1 &
    `female down-regulated` == 0 &
    `male down-regulated` == 0 &
    `sex factor down-regulated` == 0 &
    `sex factor up-regulated` == 0 & 
    `male up-regulated` == 0
  )
LBD_uniquely_down_females <-
  subset(data,
    `female down-regulated` == 1 &
    `female up-regulated` == 0 &
    `male down-regulated` == 0 &
    `sex factor down-regulated` == 0 &
    `sex factor up-regulated` == 0 & 
    `male up-regulated` == 0
  )

# combine with LBD df to get fold change information 
LBD_uniquely_up_females_df <-
  subset(LBD_FandM, gene_name %in% rownames(LBD_uniquely_up_females))

# further sort
# up in females and not also up in males, at least less than logFC of 0.025
LBD_uniquely_up_females_df_subset <-
  subset(LBD_uniquely_up_females_df, female_logFC > 0.25 & male_logFC <= 0.0)

# repeat for down regulated 
LBD_uniquely_down_females_df <-
  subset(LBD_FandM, gene_name %in% rownames(LBD_uniquely_down_females))
# further sort
LBD_uniquely_down_females_df_subset <-
  subset(LBD_uniquely_down_females_df, female_logFC < -0.25 & male_logFC >= 0)

# repeat for males
LBD_uniquely_up_males <-
  subset(data,
    `male up-regulated` == 1 &
    `female down-regulated` == 0 &
    `male down-regulated` == 0 &
    `sex factor down-regulated` == 0 &
    `sex factor up-regulated` == 0 & 
    `female up-regulated` == 0
  )
LBD_uniquely_down_males <-
  subset(data,
    `male down-regulated` == 1 &
    `female up-regulated` == 0 &
    `female down-regulated` == 0 &
    `sex factor down-regulated` == 0 &
    `sex factor up-regulated` == 0 & 
    `male up-regulated` == 0
  )

LBD_uniquely_up_males_df <-
  subset(LBD_FandM, gene_name %in% rownames(LBD_uniquely_up_males))
LBD_uniquely_up_males_df_subset <-
  subset(LBD_uniquely_up_males_df, male_logFC > 0.25 & female_logFC <= 0)

LBD_uniquely_down_males_df <-
  subset(LBD_FandM, gene_name %in% rownames(LBD_uniquely_down_males))
LBD_uniquely_down_males_df_subset <-
  subset(LBD_uniquely_down_males_df, male_logFC < -0.25 & female_logFC >= 0)

remove(data)
```

### AD vs Control bt the sexes
```{r}
list_input <- list("female down-regulated" = ADvsControl_female_XX_down$gene_name,
                   "female up-regulated" = ADvsControl_female_XX_up$gene_name,
                   "male down-regulated" = ADvsControl_male_XY_down$gene_name,
                   "male up-regulated" = ADvsControl_male_XY_up$gene_name)
data <- fromList(list_input)

upset_gene <- upset(data, set_sizes = FALSE,
      c('female down-regulated',
        'male down-regulated',
        'female up-regulated',
        'male up-regulated'),
      themes=upset_modify_themes(
        list('intersections_matrix'=theme(text=element_text(size=10),
              plot.margin = margin(0, 0, 0, 0, "cm")),
          'overall_sizes' = theme(axis.text.x = element_text(size =10), 
              plot.margin = margin(0, 0, 0, 0, "cm")))),
  queries=list(
    upset_query(set='female down-regulated', fill='blue'),
    upset_query(set='male down-regulated', fill='blue'),
    upset_query(set='female up-regulated', fill='red'),
    upset_query(set='male up-regulated', fill='red'),
    upset_query(intersect = c('female down-regulated','male down-regulated'), 
               fill='blue', only_components = 
                  c('intersections_matrix', 'Intersection size')),
    upset_query(intersect = c('female up-regulated','male up-regulated'), 
                fill='red', only_components = 
                  c('intersections_matrix', 'Intersection size'))
  ),
  intersections = list(c('female down-regulated', 'male down-regulated'), 
                       'female down-regulated', 
                       'male down-regulated',
                       c('female up-regulated', 'male up-regulated'), 
                       'female up-regulated', 
                       'male up-regulated',
                       c('female up-regulated', 'male down-regulated'), 
                       c('female down-regulated', 'male up-regulated')
  ),
  base_annotations=list(
    'Intersection size'=(
        intersection_size(
        size=2,
        text = list(size = 3),
        text_mapping = aes(),
        bar_number_threshold=3,  # show all numbers on top of bars
        width=0.3,   # reduce width of the bars
        mapping=aes(fill='bars_color')
      )
      + scale_fill_manual(values=c('grey'), guide='none')
      + scale_y_continuous(expand=expansion(mult=c(0, 0.5)))
      + theme(axis.text.y = element_text(size = 10),
              axis.title.y = element_text(size = 10),
          panel.grid.major=element_blank(),
          panel.grid.minor=element_blank(),
          plot.margin = margin(0, 0, 0, 0, "cm"), 
          axis.line=element_line(colour='black')))),
  matrix=intersection_matrix(
      geom=geom_point(
        shape='circle filled',
        size=2,
        stroke=0.45)) +
        theme(axis.text.y = element_text(size = 10),
              axis.title.y = element_text(size = 10), 
              plot.margin = margin(0, 0, 0, 0, "cm")),
    sort_sets=FALSE,
  sort_intersections=FALSE)
upset_gene

# save
path <- paste0("../../results/", tool, 
               "/upset/", condition, "_ADvsControl_bt_sexes_FDRq0.05")
saveToPDF(paste0(path, ".pdf"), width = 5, height = 4)

write.table(
  data,
  paste0("../../results/",
  tool,
  "/upset/", condition, 
  "_ADvsControl_bt_sexes_list_binary_results.txt"),
  sep = "\t",
  quote = FALSE
)
```

### AD vs Control with sex as a factor
```{r}
list_input <- list("sex factor down-regulated" = ADvsControl_down$gene_name,
                   "sex factor up-regulated" = ADvsControl_up$gene_name,
                   "female down-regulated" = ADvsControl_female_XX_down$gene_name,
                   "female up-regulated" = ADvsControl_female_XX_up$gene_name,
                   "male down-regulated" = ADvsControl_male_XY_down$gene_name,
                   "male up-regulated" = ADvsControl_male_XY_up$gene_name)
data <- fromList(list_input)

upset_gene <- upset(data, set_sizes = FALSE,
      c('sex factor down-regulated',
        'female down-regulated',
        'male down-regulated',
        'sex factor up-regulated',
        'female up-regulated',
        'male up-regulated'),
      themes=upset_modify_themes(
        list('intersections_matrix'=theme(text=element_text(size=10),
                                          plot.margin = margin(0, 0, 0, 0, "cm")),
          'overall_sizes' = theme(axis.text.x = element_text(size =10), 
                                  plot.margin = margin(0, 0, 0, 0, "cm")))),

  queries=list(
    upset_query(set='sex factor down-regulated', fill='blue'),
    upset_query(set='female down-regulated', fill='blue'),
    upset_query(set='male down-regulated', fill='blue'),
    upset_query(set='sex factor up-regulated', fill='red'),
    upset_query(set='female up-regulated', fill='red'),
    upset_query(set='male up-regulated', fill='red')
  ),
  intersections = list(c('sex factor down-regulated','female down-regulated', 'male down-regulated'), 
                       c('sex factor down-regulated', 'female down-regulated'), 
                       c('sex factor down-regulated', 'male down-regulated'), 
                       c('female down-regulated', 'male down-regulated'), 
                       'sex factor down-regulated',
                       'female down-regulated', 
                       'male down-regulated',
                       c('sex factor up-regulated','female up-regulated', 'male up-regulated'), 
                       c('sex factor up-regulated', 'female up-regulated'), 
                       c('sex factor up-regulated', 'male up-regulated'), 
                       c('female up-regulated', 'male up-regulated'), 
                       'sex factor up-regulated',
                       'female up-regulated', 
                       'male up-regulated',
                       c('female up-regulated', 'male down-regulated'), 
                       c('female down-regulated', 'male up-regulated')
  ),
  base_annotations=list(
    'Intersection size'=(
        intersection_size(
        size=2,
        text = list(size = 3),
        text_mapping = aes(),
        bar_number_threshold=3,  # show all numbers on top of bars
        width=0.3,   # reduce width of the bars
        mapping=aes(fill='bars_color')
      )
      + scale_fill_manual(values=c('grey'), guide='none')
      + scale_y_continuous(expand=expansion(mult=c(0, 0.5)))
      + theme(axis.text.y = element_text(size = 10),
              axis.title.y = element_text(size = 10),
          panel.grid.major=element_blank(),
          panel.grid.minor=element_blank(),
          plot.margin = margin(0, 0, 0, 0, "cm"), 
          axis.line=element_line(colour='black')))),
  matrix=intersection_matrix(
      geom=geom_point(
        shape='circle filled',
        size=2,
        stroke=0.45)) +
        theme(axis.text.y = element_text(size = 10),
              axis.title.y = element_text(size = 10), 
              plot.margin = margin(0, 0, 0, 0, "cm")),
    sort_sets=FALSE,
  sort_intersections=FALSE)
upset_gene

# save
path <- paste0("../../results/", tool, "/upset/", condition, "_ADvsControl_sex_factor_FDRq0.05")
saveToPDF(paste0(path, ".pdf"), width = 6, height = 4)

upset(data, intersect = c('sex factor down-regulated',
        'female down-regulated',
        'male down-regulated',
        'sex factor up-regulated',
        'female up-regulated',
        'male up-regulated'))

path <- paste0("../../results/", tool, "/upset/", condition, "_ADvsControl_sex_factor_FDRq0.05_simple")
saveToPDF(paste0(path, ".pdf"), width = 9, height = 6)

# Binary table with colnames:
write.table(
  data,
  paste0("../../results/",
  tool,
  "/upset/", condition, 
  "_ADvsControl_sex_factor_list_binary_results.txt"),
  sep = "\t",
  quote = FALSE
)
```
### AD DEGs specific to each sex 
```{r}
# merge the male and female data frames 
# this will get the logFC for the female and male only analysis
AD_FandM <-
  merge(ADvsControl_female_XX,
        ADvsControl_male_XY,
        by = "gene_id",
        all = T)
# remove unnecessary columns 
AD_FandM <- AD_FandM[c(7,1,2,6,8:15,22:29)]
colnames(AD_FandM) = c("gene_name", "gene_id","chr", "gene_type", "female_logFC", "female_CI.L", "female_CI.R", "female_AveExpr", "female_t","female_P.val", "female_adj.P.Val", "female_B", "male_logFC", "male_CI.L", "male_CI.R", "male_AveExpr", "male_t","male_P.val", "male_adj.P.Val", "male_B")
# uniquely up or down in the females 
# not DE in the male only or sex as a factor
AD_uniquely_up_females <-
  subset(data,
    `female up-regulated` == 1 &
    `female down-regulated` == 0 &
    `male down-regulated` == 0 &
    `sex factor down-regulated` == 0 &
    `sex factor up-regulated` == 0 & 
    `male up-regulated` == 0
  )
AD_uniquely_down_females <-
  subset(data,
    `female down-regulated` == 1 &
    `female up-regulated` == 0 &
    `male down-regulated` == 0 &
    `sex factor down-regulated` == 0 &
    `sex factor up-regulated` == 0 & 
    `male up-regulated` == 0
  )

# combine with AD df to get fold change information 
AD_uniquely_up_females_df <-
  subset(AD_FandM, gene_name %in% rownames(AD_uniquely_up_females))

# further sort
# up in females and not also up in males, at least less than logFC of 0.025
AD_uniquely_up_females_df_subset <-
  subset(AD_uniquely_up_females_df, female_logFC > 0.25 & male_logFC <= 0)

# repeat for down regulated 
AD_uniquely_down_females_df <-
  subset(AD_FandM, gene_name %in% rownames(AD_uniquely_down_females))
# further sort
AD_uniquely_down_females_df_subset <-
  subset(AD_uniquely_down_females_df, female_logFC < -0.25 & male_logFC >= 0)

# repeat for males
AD_uniquely_up_males <-
  subset(data,
    `male up-regulated` == 1 &
    `female down-regulated` == 0 &
    `male down-regulated` == 0 &
    `sex factor down-regulated` == 0 &
    `sex factor up-regulated` == 0 & 
    `female up-regulated` == 0
  )
AD_uniquely_down_males <-
  subset(data,
    `male down-regulated` == 1 &
    `female up-regulated` == 0 &
    `female down-regulated` == 0 &
    `sex factor down-regulated` == 0 &
    `sex factor up-regulated` == 0 & 
    `male up-regulated` == 0
  )

AD_uniquely_up_males_df <-
  subset(AD_FandM, gene_name %in% rownames(AD_uniquely_up_males))
AD_uniquely_up_males_df_subset <-
  subset(AD_uniquely_up_males_df, male_logFC > 0.25 & female_logFC <= 0)

AD_uniquely_down_males_df <-
  subset(AD_FandM, gene_name %in% rownames(AD_uniquely_down_males))
AD_uniquely_down_males_df_subset <-
  subset(AD_uniquely_down_males_df, male_logFC < -0.25 & female_logFC >= 0)

remove(data)
```

# Correlation plot
### Dummy empty plot
```{r}
LBD_FandM_goi <- subset(LBD_FandM, 
                               (female_logFC < -0.5 & male_logFC < -0.5) | # down in both 
                               (female_logFC > .5 & male_logFC > 0.5) | # up in both 
                               (female_logFC > .5 & male_logFC <= 0) | # up in females only 
                              (female_logFC <= 0 & male_logFC > 0.5)) # up in males only 
# plot
dummy_LBD_corr_plot <- ggplot(data = LBD_FandM, 
   aes(x = female_logFC, y = male_logFC, text = paste(gene_name))) +
   annotate(
    "rect",
    xmin = 0,
    xmax = 2.5,
    ymin = 2.5,
    ymax = 0,
    fill = "lightpink3",
    alpha = .5
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -2.25,
    ymin = 0,
    ymax = -2.25,
    fill = "cadetblue3",
    alpha = .5) +
  geom_abline(color = "gray40") +
 # geom_point(size = 0) + 
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
  title = "",
  x = expression(paste("XX female ", log[2](LBD/control))),
  y = expression(paste("XY male ", log[2](LBD/control))))+
  scale_y_continuous(breaks = seq(-2, 2.25, by = .5), 
                     limits = c(-2.25, 2.5), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-2, 2.25, by = .5), 
                     limits = c(-2.25, 2.5), expand = c(0,0))
 

dummy_LBD_corr_plot
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    "dummy_empty_correlation"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

```

### LBD vs control bt the sexes 
All genes
```{r}
LBD_FandM_goi <- subset(LBD_FandM, 
                               (female_logFC < -0.5 & male_logFC < -0.5) | # down in both 
                               (female_logFC > .5 & male_logFC > 0.5) | # up in both 
                               (female_logFC > .5 & male_logFC <= 0) | # up in females only 
                              (female_logFC <= 0 & male_logFC > 0.5)) # up in males only 
# plot
LBD_corr_plot_adjPval_1 <- ggplot(data = LBD_FandM, 
   aes(x = female_logFC, y = male_logFC, text = paste(gene_name))) +
   annotate(
    "rect",
    xmin = 0,
    xmax = 2,
    ymin = 2,
    ymax = 0,
    fill = "lightpink3",
    alpha = .5
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -2,
    ymin = 0,
    ymax = -2,
    fill = "cadetblue3",
    alpha = .5) +
  geom_abline(color = "gray40") +
    geom_text_repel(
    data = LBD_FandM_goi, 
    aes(
    x = female_logFC, 
    y = male_logFC,
    label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 20)
    ) +
  geom_point(size = 1) + 
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
  title = "All genes",
  x = expression(paste("XX female ", log[2](LBD/control))),
  y = expression(paste("XY male ", log[2](LBD/control))))+
  scale_y_continuous(breaks = seq(-2, 2, by = .5), 
                     limits = c(-2, 2), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-2, 2, by = .5), 
                     limits = c(-2, 2), expand = c(0,0))
LBD_corr_plot_adjPval_1 

LBD_corr_plot_adjPval_1
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_LBDvsControl_bt_sexes"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(LBD_FandM, x = "female_logFC", y = "male_logFC",
   add = "reg.line",  # Add regressin line
   add.params = list(color = "blue", fill = "lightgray"), # Customize reg. line
   conf.int = TRUE # Add confidence interval
   )
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -1.25, label.y = 1)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_LBDvsControl_bt_sexes_corr_pval"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)
```
significant in either sex 

```{r}
LBD_FandM <- subset(LBD_FandM, (female_adj.P.Val <= 0.05 & female_logFC > 0.25) | 
                      (female_adj.P.Val <= 0.05 & female_logFC < -0.25) |
                    (male_adj.P.Val <= 0.05 & male_logFC > 0.25) |
                    (male_adj.P.Val <= 0.05 & male_logFC < -0.25))
LBD_FandM_goi <- subset(LBD_FandM, 
                               (female_logFC < -0.5 & male_logFC < -0.5) | # down in both 
                               (female_logFC > .5 & male_logFC > 0.5) | # up in both 
                               (female_logFC > .5 & male_logFC <= 0) | # up in females only 
                              (female_logFC <= 0 & male_logFC > 0.5)) # up in males only 

sig_f <- subset(LBD_FandM, female_adj.P.Val <= 0.05 & male_adj.P.Val > 0.05)
sig_m <- subset(LBD_FandM, male_adj.P.Val <= 0.05 & female_adj.P.Val > 0.05)
sig_both <- subset(LBD_FandM, male_adj.P.Val <= 0.05 & female_adj.P.Val <= 0.05)

sig_f_sub <- subset(sig_f, female_logFC > 1 | female_logFC < -1)
sig_m_sub <- subset(sig_m, male_logFC > 1 | male_logFC < -1)
sig_both_sub <- subset(sig_both, male_logFC > 1 | male_logFC < -1 | female_logFC > 1 | female_logFC < -1)

# plot
LBD_corr_plot_adjPval_05 <- ggplot(data = LBD_FandM, 
   aes(x = female_logFC, y = male_logFC, text = paste(gene_name))) +
   annotate(
    "rect",
    xmin = 0,
    xmax = 2,
    ymin = 2,
    ymax = 0,
    fill = "lightpink3",
    alpha = .1
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -2,
    ymin = 0,
    ymax = -2,
    fill = "cadetblue3",
    alpha = .1) +
  geom_abline(color = "gray80", linetype = 2) +
  geom_point(data = sig_f, color="#490092", shape=16) + 
  geom_point(data = sig_m, color="#D55E00", shape=15) + 
  geom_point(data = sig_both, color="gray20", shape=25) + 
  geom_text_repel(
    data = sig_both_sub, 
    aes(
    x = female_logFC, 
    y = male_logFC,
    label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 18)
    ) +
  geom_text_repel(
    data = sig_f_sub, 
    aes(
    x = female_logFC, 
    y = male_logFC,
    label = gene_name
    ),
    color = "#490092",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 18)
    ) +
    geom_text_repel(
    data = sig_m_sub, 
    aes(
    x = female_logFC, 
    y = male_logFC,
    label = gene_name
    ),
    color = "#D55E00",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 15)
    ) +
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
  title = "DEGs shared and unique between the sexes
adj.P.val <= 0.05 & absolute log2FC > 0.25",
  x = expression(paste("XX female ", log[2](LBD/control))),
  y = expression(paste("XY male ", log[2](LBD/control))))+
  scale_y_continuous(breaks = seq(-2, 2, by = .5), 
                     limits = c(-2, 2), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-2, 2, by = .5), 
                     limits = c(-2, 2), expand = c(0,0))
LBD_corr_plot_adjPval_05 

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_LBDvsControl_bt_sexes_sig"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(LBD_FandM, x = "female_logFC", y = "male_logFC",
   add = "reg.line",  # Add regressin line
   add.params = list(color = "blue", fill = "lightgray"), # Customize reg. line
   conf.int = TRUE # Add confidence interval
   )
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -1.25, label.y = 1)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_LBDvsControl_bt_sexes_sig_corr_pval"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)
```

Significant in only XX females
```{r}
LBD_FandM <- subset(LBD_FandM, (female_adj.P.Val <= 0.05 & female_logFC > 0.25) | 
                      (female_adj.P.Val <= 0.05 & female_logFC < -0.25) |
                    (male_adj.P.Val <= 0.05 & male_logFC > 0.25) |
                    (male_adj.P.Val <= 0.05 & male_logFC < -0.25))
LBD_FandM_goi <- subset(LBD_FandM, 
                               (female_logFC < -0.5 & male_logFC < -0.5) | # down in both 
                               (female_logFC > .5 & male_logFC > 0.5) | # up in both 
                               (female_logFC > .5 & male_logFC <= 0) | # up in females only 
                              (female_logFC <= 0 & male_logFC > 0.5)) # up in males only 

sig_f <- subset(LBD_FandM, female_adj.P.Val <= 0.05 & male_adj.P.Val > 0.05)
sig_m <- subset(LBD_FandM, male_adj.P.Val <= 0.05 & female_adj.P.Val > 0.05)
sig_both <- subset(LBD_FandM, male_adj.P.Val <= 0.05 & female_adj.P.Val <= 0.05)

sig_f_sub <- subset(sig_f, female_logFC > 1 | female_logFC < -1)
sig_m_sub <- subset(sig_m, male_logFC > 1 | male_logFC < -1)
sig_both_sub <- subset(sig_both, male_logFC > 1 | male_logFC < -1 | female_logFC > 1 | female_logFC < -1)

# plot
LBD_corr_plot_adjPval_05 <- ggplot(data = LBD_FandM, 
   aes(x = female_logFC, y = male_logFC, text = paste(gene_name))) +
   annotate(
    "rect",
    xmin = 0,
    xmax = 2,
    ymin = 2,
    ymax = 0,
    fill = "lightpink3",
    alpha = .1
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -2,
    ymin = 0,
    ymax = -2,
    fill = "cadetblue3",
    alpha = .1) +
  geom_abline(color = "gray80", linetype = 2) +
  geom_point(data = sig_f, color="#490092", shape=16) + 
#  geom_point(data = sig_m, color="#D55E00", shape=15) + 
#  geom_point(data = sig_both, color="gray20", shape=25) + 
  geom_text_repel(
    data = sig_f_sub, 
    aes(
    x = female_logFC, 
    y = male_logFC,
    label = gene_name
    ),
    color = "#490092",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 20)
    ) +
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
  title = "DEGs unique to XX females
adj.P.val <= 0.05 & absolute log2FC > 0.25",
  x = expression(paste("XX female ", log[2](LBD/control))),
  y = expression(paste("XY male ", log[2](LBD/control))))+
  scale_y_continuous(breaks = seq(-2, 2, by = .5), 
                     limits = c(-2, 2), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-2, 2, by = .5), 
                     limits = c(-2, 2), expand = c(0,0))
LBD_corr_plot_adjPval_05 

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_LBDvsControl_bt_sexes_sig_only_in_XX_females"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)
```
Significant in only XY males
```{r}
LBD_FandM <- subset(LBD_FandM, (female_adj.P.Val <= 0.05 & female_logFC > 0.25) | 
                      (female_adj.P.Val <= 0.05 & female_logFC < -0.25) |
                    (male_adj.P.Val <= 0.05 & male_logFC > 0.25) |
                    (male_adj.P.Val <= 0.05 & male_logFC < -0.25))
LBD_FandM_goi <- subset(LBD_FandM, 
                               (female_logFC < -0.5 & male_logFC < -0.5) | # down in both 
                               (female_logFC > .5 & male_logFC > 0.5) | # up in both 
                               (female_logFC > .5 & male_logFC <= 0) | # up in females only 
                              (female_logFC <= 0 & male_logFC > 0.5)) # up in males only 

sig_f <- subset(LBD_FandM, female_adj.P.Val <= 0.05 & male_adj.P.Val > 0.05)
sig_m <- subset(LBD_FandM, male_adj.P.Val <= 0.05 & female_adj.P.Val > 0.05)
sig_both <- subset(LBD_FandM, male_adj.P.Val <= 0.05 & female_adj.P.Val <= 0.05)

sig_f_sub <- subset(sig_f, female_logFC > 1 | female_logFC < -1)
sig_m_sub <- subset(sig_m, male_logFC > .25 | male_logFC < -.25)
sig_both_sub <- subset(sig_both, male_logFC > 1 | male_logFC < -1 | female_logFC > 1 | female_logFC < -1)

# plot
LBD_corr_plot_adjPval_05 <- ggplot(data = LBD_FandM, 
   aes(x = female_logFC, y = male_logFC, text = paste(gene_name))) +
   annotate(
    "rect",
    xmin = 0,
    xmax = 2,
    ymin = 2,
    ymax = 0,
    fill = "lightpink3",
    alpha = .1
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -2,
    ymin = 0,
    ymax = -2,
    fill = "cadetblue3",
    alpha = .1) +
  geom_abline(color = "gray80", linetype = 2) +
  geom_point(data = sig_m, color="#D55E00", shape=15) + 
    geom_text_repel(
    data = sig_m_sub, 
    aes(
    x = female_logFC, 
    y = male_logFC,
    label = gene_name
    ),
    color = "#D55E00",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 20)
    ) +
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
  title = "DEGs unique to XY males
adj.P.val <= 0.05 & absolute log2FC > 0.25",
  x = expression(paste("XX female ", log[2](LBD/control))),
  y = expression(paste("XY male ", log[2](LBD/control))))+
  scale_y_continuous(breaks = seq(-2, 2, by = .5), 
                     limits = c(-2, 2), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-2, 2, by = .5), 
                     limits = c(-2, 2), expand = c(0,0))
LBD_corr_plot_adjPval_05 

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_LBDvsControl_bt_sexes_sig_only_in_XY_males"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)
```

### AD vs control bt the sexes 
```{r}
AD_FandM_goi <- subset(AD_FandM, 
                               (female_logFC < -0.5 & male_logFC < -0.5) | # down in both 
                               (female_logFC > .5 & male_logFC > 0.5) | # up in both 
                               (female_logFC > .5 & male_logFC <= 0) | # up in females only 
                              (female_logFC <= 0 & male_logFC > 0.5)) # up in males only 
# plot
AD_corr_plot_adjPval_1 <- ggplot(data = AD_FandM, 
   aes(x = female_logFC, y = male_logFC, text = paste(gene_name))) +
   annotate(
    "rect",
    xmin = 0,
    xmax = 2.5,
    ymin = 2.5,
    ymax = 0,
    fill = "lightpink3",
    alpha = .5
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -2.25,
    ymin = 0,
    ymax = -2.25,
    fill = "cadetblue3",
    alpha = .5) +
  geom_abline(color = "gray40") +
    geom_text_repel(
    data = AD_FandM_goi, 
    aes(
    x = female_logFC, 
    y = male_logFC,
    label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 15)
    ) +
  geom_point(size = 1) + 
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
  title = "All genes",
  x = expression(paste("XX female ", log[2](AD/control))),
  y = expression(paste("XY male ", log[2](AD/control))))+
  scale_y_continuous(breaks = seq(-2, 2.25, by = .5), 
                     limits = c(-2.25, 2.5), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-2, 2.5, by = .5), 
                     limits = c(-2.25, 2.5), expand = c(0,0))
AD_corr_plot_adjPval_1 

AD_corr_plot_adjPval_1
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_ADvsControl_bt_sexes"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(AD_FandM, x = "female_logFC", y = "male_logFC",
   add = "reg.line",  # Add regressin line
   add.params = list(color = "blue", fill = "lightgray"), # Customize reg. line
   conf.int = TRUE # Add confidence interval
   )
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -1.25, label.y = 1)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_ADvsControl_bt_sexes_corr_pval"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# significant only 
AD_FandM <- subset(AD_FandM, (female_adj.P.Val <= 0.05 & female_logFC > 0.25) | 
                      (female_adj.P.Val <= 0.05 & female_logFC < -0.25) |
                    (male_adj.P.Val <= 0.05 & male_logFC > 0.25) |
                    (male_adj.P.Val <= 0.05 & male_logFC < -0.25))
AD_FandM_goi <- subset(AD_FandM, 
                               (female_logFC < -0.5 & male_logFC < -0.5) | # down in both 
                               (female_logFC > .5 & male_logFC > 0.5) | # up in both 
                               (female_logFC > .5 & male_logFC <= 0) | # up in females only 
                              (female_logFC <= 0 & male_logFC > 0.5)) # up in males only 

sig_f <- subset(AD_FandM, female_adj.P.Val <= 0.05 & male_adj.P.Val > 0.05)
sig_m <- subset(AD_FandM, male_adj.P.Val <= 0.05 & female_adj.P.Val > 0.05)
sig_both <- subset(AD_FandM, male_adj.P.Val <= 0.05 & female_adj.P.Val <= 0.05)

sig_f_sub <- subset(sig_f, female_logFC > 1 | female_logFC < -1)
sig_m_sub <- subset(sig_m, male_logFC > 1 | male_logFC < -1)
sig_both_sub <- subset(sig_both, male_logFC > 1 | male_logFC < -1 | female_logFC > 1 | female_logFC < -1)

# plot
AD_corr_plot_adjPval_05 <- ggplot(data = AD_FandM, 
   aes(x = female_logFC, y = male_logFC, text = paste(gene_name))) +
   annotate(
    "rect",
    xmin = 0,
    xmax = 2.5,
    ymin = 2.5,
    ymax = 0,
    fill = "lightpink3",
    alpha = .1
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -2.25,
    ymin = 0,
    ymax = -2.25,
    fill = "cadetblue3",
    alpha = .1) +
  geom_abline(color = "gray80", linetype = 2) +
  geom_point(data = sig_f, color="#490092", shape=16) + 
  geom_point(data = sig_m, color="#D55E00", shape=15) + 
  geom_point(data = sig_both, color="gray20", shape=25) + 
  geom_text_repel(
    data = sig_both_sub, 
    aes(
    x = female_logFC, 
    y = male_logFC,
    label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 15)
    ) +
  geom_text_repel(
    data = sig_f_sub, 
    aes(
    x = female_logFC, 
    y = male_logFC,
    label = gene_name
    ),
    color = "#490092",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 15)
    ) +
    geom_text_repel(
    data = sig_m_sub, 
    aes(
    x = female_logFC, 
    y = male_logFC,
    label = gene_name
    ),
    color = "#D55E00",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 15)
    ) +
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
  title = "DEGs shared and unique between the sexes
adj.P.val <= 0.05 & absolute log2FC > 0.25",
  x = expression(paste("XX female ", log[2](AD/control))),
  y = expression(paste("XY male ", log[2](AD/control))))+
  scale_y_continuous(breaks = seq(-2, 2.25, by = .5), 
                     limits = c(-2.25, 2.5), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-2, 2.5, by = .5), 
                     limits = c(-2.25, 2.5), expand = c(0,0))
AD_corr_plot_adjPval_05 

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_ADvsControl_bt_sexes_sig"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(AD_FandM, x = "female_logFC", y = "male_logFC",
   add = "reg.line",  # Add regressin line
   add.params = list(color = "blue", fill = "lightgray"), # Customize reg. line
   conf.int = TRUE # Add confidence interval
   )
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -1.25, label.y = 1)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_ADvsControl_bt_sexes_sig_corr_pval"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)
```
### LBD vs AD 
##### XX females
```{r}
ADandLBD_female_XX <-
  merge(ADvsControl_female_XX,
        LBDvsControl_female_XX,
        by = "gene_name",
        all = T)
ADandLBD_female_XX <- ADandLBD_female_XX[c(1,2,6,7,8:15,22:29)]
colnames(ADandLBD_female_XX) = c("gene_name", "chr", "gene_id","gene_type", "AD_female_logFC", "AD_female_CI.L", "AD_female_CI.R", "AD_female_AveExpr", "AD_female_t","AD_female_P.val", "AD_female_adj.P.Val", "AD_female_B", "LBD_female_logFC", "LBD_female_CI.L", "LBD_female_CI.R", "LBD_female_AveExpr", "LBD_female_t","LBD_female_P.val", "LBD_female_adj.P.Val", "LBD_female_B")

ADandLBD_female_XX_goi <- subset(ADandLBD_female_XX, 
                        (AD_female_logFC < -0.5 & LBD_female_logFC < -0.5) | # down in both
                          (AD_female_logFC > .5 & LBD_female_logFC > 0.5) | # up in both
                          (AD_female_logFC > .5 & LBD_female_logFC < 0) | # up in AD, down in LBD
                          (AD_female_logFC < -.5 & LBD_female_logFC > 0)) # up in LBD, down in AD

ADandLBD_female_XX_corr_plot <- ggplot(data = ADandLBD_female_XX, 
  aes(x = AD_female_logFC, y = LBD_female_logFC, text = paste(gene_name))) +
  annotate(
    "rect",
    xmin = 0,
    xmax = 2.75,
    ymin = 2.75,
    ymax = 0,
    fill = "lightpink3",
    alpha = .5
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -2.25,
    ymin = 0,
    ymax = -2.25,
    fill = "cadetblue3",
    alpha = .5) +
  geom_abline(color = "gray40") +
  geom_text_repel(
    data = ADandLBD_female_XX_goi, 
    aes(
      x = AD_female_logFC, 
      y = LBD_female_logFC,
      label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 40)
  ) +
  geom_point(size = 1) + 
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
    title = "DEGs shared and unique between LBD and AD within XX females",
    x = expression(paste("XX female ", log[2](AD/control))),
    y = expression(paste("XX female ", log[2](LBD/control))))+
  scale_y_continuous(breaks = seq(-1.75, 2.725, by = .5), 
                     limits = c(-2.25, 2.75), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-2, 2.75, by = .5), 
                     limits = c(-2.25, 2.75), expand = c(0,0))
ADandLBD_female_XX_corr_plot

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_ADandLBD_female_XX"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(ADandLBD_female_XX, x = "AD_female_logFC", y = "LBD_female_logFC",
                add = "reg.line",  # Add regression line
                add.params = list(color = "blue", fill = "lightgray"), 
                conf.int = TRUE)
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -1.25, label.y = 1)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_ADandLBD_female_XX_corr_pval"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)
```
#### XY males
```{r}
ADandLBD_male_XY <-
  merge(ADvsControl_male_XY,
        LBDvsControl_male_XY,
        by = "gene_name",
        all = T)
ADandLBD_male_XY <- ADandLBD_male_XY[c(1,2,6,7,8:15,22:29)]
colnames(ADandLBD_male_XY) = c("gene_name", "chr", "gene_id","gene_type", "AD_male_logFC", "AD_male_CI.L", "AD_male_CI.R", "AD_male_AveExpr", "AD_male_t","AD_male_P.val", "AD_male_adj.P.Val", "AD_male_B", "LBD_male_logFC", "LBD_male_CI.L", "LBD_male_CI.R", "LBD_male_AveExpr", "LBD_male_t","LBD_male_P.val", "LBD_male_adj.P.Val", "LBD_male_B")

ADandLBD_male_XY_goi <- subset(ADandLBD_male_XY, 
                        (AD_male_logFC < -0.5 & LBD_male_logFC < -0.5) | # down in both
                          (AD_male_logFC > .5 & LBD_male_logFC > 0.5) | # up in both
                          (AD_male_logFC > .5 & LBD_male_logFC < 0) | # up in AD, down in LBD
                          (AD_male_logFC < -.5 & LBD_male_logFC > 0)) # up in LBD, down in AD

ADandLBD_male_XY_corr_plot <- ggplot(data = ADandLBD_male_XY, 
  aes(x = AD_male_logFC, y = LBD_male_logFC, text = paste(gene_name))) +
  annotate(
    "rect",
    xmin = 0,
    xmax = 2.75,
    ymin = 2.75,
    ymax = 0,
    fill = "lightpink3",
    alpha = .5
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -2.25,
    ymin = 0,
    ymax = -2.25,
    fill = "cadetblue3",
    alpha = .5) +
  geom_abline(color = "gray40") +
  geom_text_repel(
    data = ADandLBD_male_XY_goi, 
    aes(
      x = AD_male_logFC, 
      y = LBD_male_logFC,
      label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 40)
  ) +
  geom_point(size = 1) + 
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
    title = "DEGs shared and unique between LBD and AD within XY males",
    x = expression(paste("XY male ", log[2](AD/control))),
    y = expression(paste("XY male ", log[2](LBD/control))))+
  scale_y_continuous(breaks = seq(-1.75, 2.725, by = .5), 
                     limits = c(-2.25, 2.75), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-2, 2.75, by = .5), 
                     limits = c(-2.25, 2.75), expand = c(0,0))
ADandLBD_male_XY_corr_plot

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_ADandLBD_male_XY"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(ADandLBD_male_XY, x = "AD_male_logFC", y = "LBD_male_logFC",
                add = "reg.line",  # Add regression line
                add.params = list(color = "blue", fill = "lightgray"), 
                conf.int = TRUE)
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -1.25, label.y = 1)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_ADandLBD_male_XY_corr_pval"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)
```
#### Sex as a factor
```{r}
ADandLBD <-
  merge(ADvsControl,
        LBDvsControl,
        by = "gene_name",
        all = T)
ADandLBD <- ADandLBD[c(1,2,6,7,8:15,22:29)]
colnames(ADandLBD) = c("gene_name", "chr", "gene_id","gene_type", "AD_logFC", "AD_CI.L", "AD_CI.R", "AD_AveExpr", "AD_t","AD_P.val", "AD_adj.P.Val", "AD_B", "LBD_logFC", "LBD_CI.L", "LBD_CI.R", "LBD_AveExpr", "LBD_t","LBD_P.val", "LBD_adj.P.Val", "LBD_B")

ADandLBD_goi <- subset(ADandLBD, 
                        (AD_logFC < -0.5 & LBD_logFC < -0.5) | # down in both
                          (AD_logFC > .5 & LBD_logFC > 0.5) | # up in both
                          (AD_logFC > .5 & LBD_logFC < 0) | # up in AD, down in LBD
                          (AD_logFC < -.5 & LBD_logFC > 0)) # up in LBD, down in AD

ADandLBD_corr_plot <- ggplot(data = ADandLBD, 
  aes(x = AD_logFC, y = LBD_logFC, text = paste(gene_name))) +
  annotate(
    "rect",
    xmin = 0,
    xmax = 2.75,
    ymin = 2.75,
    ymax = 0,
    fill = "lightpink3",
    alpha = .5
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -2.25,
    ymin = 0,
    ymax = -2.25,
    fill = "cadetblue3",
    alpha = .5) +
  geom_abline(color = "gray40") +
  geom_text_repel(
    data = ADandLBD_goi, 
    aes(
      x = AD_logFC, 
      y = LBD_logFC,
      label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 40)
  ) +
  geom_point(size = 1) + 
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
    title = "DEGs shared and unique between LBD and AD",
    x = expression(paste("  ", log[2](AD/control))),
    y = expression(paste("  ", log[2](LBD/control))))+
  scale_y_continuous(breaks = seq(-1.75, 2.725, by = .5), 
                     limits = c(-2.25, 2.75), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-2, 2.75, by = .5), 
                     limits = c(-2.25, 2.75), expand = c(0,0))
ADandLBD_corr_plot

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_ADandLBD_sex_correction"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(ADandLBD, x = "AD_logFC", y = "LBD_logFC",
                add = "reg.line",  # Add regression line
                add.params = list(color = "blue", fill = "lightgray"), 
                conf.int = TRUE)
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -1.25, label.y = 1)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/", condition, 
    "_ADandLBD_sex_correction_corr_pval"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)
```
# Heatmaps
```{r}
LBD_FandM <-
  merge(LBDvsControl_female_XX,
        LBDvsControl_male_XY,
        by = "gene_id",
        all = T)
# remove unnecessary columns 
LBD_FandM <- LBD_FandM[c(7,1,2,6,8:15,22:29)]
colnames(LBD_FandM) = c("gene_name", "gene_id","chr", "gene_type", "female_logFC", "female_CI.L", "female_CI.R", "female_AveExpr", "female_t","female_P.val", "female_adj.P.Val", "female_B", "male_logFC", "male_CI.L", "male_CI.R", "male_AveExpr", "male_t","male_P.val", "male_adj.P.Val", "male_B")

LBD_FandM$difference <- LBD_FandM$female_logFC - LBD_FandM$male_logFC
ATEST <- merge(LBD_FandM, LBDvsControl_sex_interaction, by = "gene_id", all = T)

LBD_FandM <- subset(LBD_FandM, difference > .75 | difference < -.75)
LBD_FandM <- LBD_FandM[order(-LBD_FandM$difference),]
LBD_FandM_DEGs_melt <- reshape2::melt(LBD_FandM)
data <- subset(LBD_FandM_DEGs_melt, variable == "female_logFC" | variable == "male_logFC")
data$gene_name <- factor(data$gene_name , levels = unique(data$gene_name))
data$gene_name <- fct_rev(data$gene_name)

ggplot(data = data) +
  geom_tile(aes(x = variable, y =gene_name, fill = value)) +
  scale_fill_gradient2(
    low = "blue",
    mid = "white",
    high = "red",
    midpoint = 0,
    space = "rgb",
    guide = "colourbar",
    breaks = c(-2.5, 0, 2.5),
    name = expression(log[2](CPM))
  ) +
  scale_x_discrete(expand = c(0,0)) +
  theme_bw() +
  theme(
    strip.text = element_text(size = 10),
    legend.position = "none",
    panel.border = element_blank(),
    panel.background = element_blank(),
    plot.margin = margin(0, 0.2, 0, 0.2, "cm"), 
    panel.spacing = unit(0,'lines'), 
    plot.title = element_text(size = 10, vjust = -1, hjust = 0),
    axis.title.x=element_blank(),
   # axis.text.x = element_blank(), 
    axis.ticks.x=element_blank(), 
    axis.title.y = element_blank(), 
    axis.text.y = element_text(size = 10, margin = margin(r = -2)),
    axis.ticks.y=element_blank()) +
  ggtitle("Genes with the largest
difference in log2FC between the sexes") 

path <-
  paste0(
    "../../results/",
    tool,
    "/heatmap/", condition, 
    "_LBDvsControl_log2FC_bt_sexes"
  )
saveToPDF(paste0(path, ".pdf"), width = 4, height = 9)
```

```{r}
AD_FandM <-
  merge(ADvsControl_female_XX,
        ADvsControl_male_XY,
        by = "gene_id",
        all = T)
# remove unnecessary columns 
AD_FandM <- AD_FandM[c(7,1,2,6,8:15,22:29)]
colnames(AD_FandM) = c("gene_name", "gene_id","chr", "gene_type", "female_logFC", "female_CI.L", "female_CI.R", "female_AveExpr", "female_t","female_P.val", "female_adj.P.Val", "female_B", "male_logFC", "male_CI.L", "male_CI.R", "male_AveExpr", "male_t","male_P.val", "male_adj.P.Val", "male_B")

AD_FandM$difference <- AD_FandM$female_logFC - AD_FandM$male_logFC
ATEST <- merge(AD_FandM, ADvsControl_sex_interaction, by = "gene_id", all = T)

AD_FandM <- subset(AD_FandM, difference > .75 | difference < -.75)
AD_FandM <- AD_FandM[order(-AD_FandM$difference),]
AD_FandM_DEGs_melt <- reshape2::melt(AD_FandM)
data <- subset(AD_FandM_DEGs_melt, variable == "female_logFC" | variable == "male_logFC")
data$gene_name <- factor(data$gene_name , levels = unique(data$gene_name))
data$gene_name <- fct_rev(data$gene_name)

ggplot(data = data) +
  geom_tile(aes(x = variable, y =gene_name, fill = value)) +
  scale_fill_gradient2(
    low = "blue",
    mid = "white",
    high = "red",
    midpoint = 0,
    space = "rgb",
    guide = "colourbar",
    breaks = c(-2.5, 0, 2.5),
    name = expression(log[2](CPM))
  ) +
  scale_x_discrete(expand = c(0,0)) +
  theme_bw() +
  theme(
    strip.text = element_text(size = 10),
    legend.position = "none",
    panel.border = element_blank(),
    panel.background = element_blank(),
    plot.margin = margin(0, 0.2, 0, 0.2, "cm"), 
    panel.spacing = unit(0,'lines'), 
    plot.title = element_text(size = 10, vjust = -1, hjust = 0),
    axis.title.x=element_blank(),
   # axis.text.x = element_blank(), 
    axis.ticks.x=element_blank(), 
    axis.title.y = element_blank(), 
    axis.text.y = element_text(size = 10, margin = margin(r = -2)),
    axis.ticks.y=element_blank()) +
  ggtitle("Genes with the largest
difference in log2FC between the sexes") 

path <-
  paste0(
    "../../results/",
    tool,
    "/heatmap/", condition, 
    "_ADvsControl_log2FC_bt_sexes"
  )
saveToPDF(paste0(path, ".pdf"), width = 4, height = 20)
```

# Bar plot
```{r}
# Use dose as a factor rather than numeric
df_LBD <- LBDvsControl_sex_interaction[order(LBDvsControl_sex_interaction$logFC),]

#subset data to keep confident differences 
df_LBD <- subset(df_LBD, (logFC >= .5 & CI.L > 0.1) | (logFC <= -.5 & CI.R < -0.25))

# lock in factor level order
df_LBD$gene_name <- factor(df_LBD$gene_name, levels = df_LBD$gene_name[order(df_LBD$logFC)])
# Error bars represent standard error of the mean
ggplot(df_LBD, aes(x=logFC, y=gene_name)) + 
    geom_bar(position=position_dodge(), stat="identity") +
    geom_errorbar(aes(xmin=CI.L, xmax=CI.R),
                  width=.2,                    # Width of the error bars
                  position=position_dodge(.9)) +
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10),
        axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
    title = "Genes with the largest difference between the sexes",
    x = "log2FC difference between females and males",
    y = "") +
  geom_vline(xintercept = 0, color = "blue", size=.5)

path <-
  paste0(
    "../../results/",
    tool,
    "/bar/", condition, 
    "_LBDvsControl_log2FC_bt_sexes"
  )
saveToPDF(paste0(path, ".pdf"), width = 8, height = 9)
```

```{r}
df_AD <- ADvsControl_sex_interaction[order(ADvsControl_sex_interaction$logFC),]

#subset data to keep confident differences 
df_AD <- subset(df_AD, (logFC >= .5 & CI.L > 0.01) | (logFC <= -.5 & CI.R < -0.01))

# lock in factor level order
df_AD$gene_name <- factor(df_AD$gene_name, levels = df_AD$gene_name)
# Error bars represent standard error of the mean
ggplot(df_AD, aes(x=logFC, y=gene_name)) + 
    geom_bar(position=position_dodge(), stat="identity") +
    geom_errorbar(aes(xmin=CI.L, xmax=CI.R),
                  width=.2,                    # Width of the error bars
                  position=position_dodge(.9)) +
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10),
        axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
    title = "Genes with the largest difference between the sexes",
    x = "log2FC difference between females and males",
    y = "") +
  geom_vline(xintercept = 0, color = "blue", size=.5)

path <-
  paste0(
    "../../results/",
    tool,
    "/bar/", condition, 
    "_ADvsControl_log2FC_bt_sexes"
  )
saveToPDF(paste0(path, ".pdf"), width = 5, height = 25)
```

```{r}
```
