==> scripts/R/08_compare_DEGs_to_other_datasets.Rmd <==
---
title: "Compare DEGs to other datasets"
author: "Kimberly Olney"
date: "01/06/2023"
output: html_document
---

# Setup
```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = ".")
```
# User defined variables
```{r set_variables}
source(here::here("scripts/R", "file_paths_and_colours.R"))
condition <- c("TYPE")
library(plotly)
tool <- "star"
# save contrast names
allComparisons <- c("LBDvsControl", 
                    "LBDvsAD", 
                    "LBDvsPA", 
                    "ADvsControl", 
                    "PAvsControl", 
                    "ADvsPA")
allComparisons # check
```
# Read in DEGs
Read table with all genes (FDRq = 1).
```{r read_DEG_table}
# sex as a covariate
coef <- 1
for (i in allComparisons) {
  filepath <- paste0(
    "../../results/",
    tool,
    "/DEGs/", condition, "_", 
    i,
    "_gene_DEGs_FDRq1.00.txt"
  )
  assign(paste0(i),
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ))
  # increment
  coef <- coef + 1
}
```

```{r}
#test <- vroom("/research/labs/neurology/fryer/m239830/LBD_CWOW/other_datasets/Marques-Coelho_2021/sup_table_1.txt")
#colnames(test)[4] ="gene_name"

ROSMAP_AD <- read.delim("/research/labs/neurology/fryer/m239830/synapse_RNAseq_harmonization/ROSMAP_Differential_Expression_diagnosis_PCC.tsv", sep = "\t", header = TRUE)

ROSMAP_AD$gene_name <- ROSMAP_AD$hgnc_symbol
```
# CWOW vs ROSMAP - AD 
```{r}
data <-
  merge(ADvsControl,
        ROSMAP_AD,
        by = "gene_name",
        all = T)
data <- data[c(1,2,6,8,11,14,18,21,24)] #data[c(1,2,6,9,10,13,22,23,26)]
colnames(data) = c("gene_name", "chr", "gene_id", "AD_CWOW_logFC", "AD_CWOW_AveExpr", "AD_CWOW_adj.P.Val", "AD_ROSMAP_logFC", "AD_ROSMAP_AveExpr", "AD_ROSMAP_adj.P.Val")

#data[is.na(data)] <- 0
data <- na.omit(data)
CWOW <- subset(data, AD_CWOW_adj.P.Val <= 0.05)
ROSMAP <- subset(data, AD_ROSMAP_adj.P.Val <= 0.05) 

data <- subset(data, AD_CWOW_adj.P.Val <= 0.05 & AD_ROSMAP_adj.P.Val <= 0.05)
data <- subset(data, AD_CWOW_logFC <= -0.25 | AD_CWOW_logFC >= 0.25)
data <- subset(data, AD_ROSMAP_logFC <= -0.25 | AD_ROSMAP_logFC >= 0.25)

data_goi <- subset(data, 
                        (AD_ROSMAP_logFC < -0.5 & AD_CWOW_logFC < -0.75) | 
                          (AD_ROSMAP_logFC > .5 & AD_CWOW_logFC > 0.5) | 
                          (AD_ROSMAP_logFC > .25 & AD_CWOW_logFC <= 0) |
                          (AD_ROSMAP_logFC <= 0 & AD_CWOW_logFC > 0.25))

#data$difference <- data$logFC - data$gene.log2FC
ggplot(data = data, 
  aes(x = AD_ROSMAP_logFC, y = AD_CWOW_logFC, text = paste(gene_name))) +
  annotate(
    "rect",
    xmin = 0,
    xmax = 1.75,
    ymin = 1.75,
    ymax = 0,
    fill = "lightpink3",
    alpha = .5
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -1.5,
    ymin = 0,
    ymax = -1.5,
    fill = "cadetblue3",
    alpha = .5) +
  geom_abline(color = "gray40") +
  geom_text_repel(
    data = data_goi, 
    aes(
      x = AD_ROSMAP_logFC, 
      y = AD_CWOW_logFC,
      label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 10)
  ) +
  geom_point(size = 1) + 
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
    title = "DEGs shared and unique between LBD CWOW AD and ROSMAP PCC",
    x = expression(paste("ROSMAP ", log[2](AD/control))),
    y = expression(paste("LBD CWOW ", log[2](AD/control))))+
  scale_y_continuous(breaks = seq(-1.5, 1.75, by = .5), 
                     limits = c(-1.5, 1.75), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-1.5, 1.75, by = .5), 
                     limits = c(-1.5, 1.75), expand = c(0,0))

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    condition, 
    "_AD_CWOW_vs_ROSMAP_FDRq0.05_log2FC0.25"
  )
#saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(data, x = "AD_ROSMAP_logFC", y = "AD_CWOW_logFC",
                add = "reg.line",  # Add regressin line
                add.params = list(color = "blue", fill = "lightgray"), 
                conf.int = TRUE)
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -.5, label.y = .75)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    condition, 
    "_AD_CWOW_vs_ROSMAP_FDRq0.05_log2FC0.25_corr_pval"
  )
#saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)
```
```{r}
data <-
  merge(ADvsControl,
        ROSMAP_AD,
        by = "gene_name",
        all = T)
data <- data[c(1,2,6,8,11,14,18,21,24)] #data[c(1,2,6,9,10,13,22,23,26)]
colnames(data) = c("gene_name", "chr", "gene_id", "AD_CWOW_logFC", "AD_CWOW_AveExpr", "AD_CWOW_adj.P.Val", "AD_ROSMAP_logFC", "AD_ROSMAP_AveExpr", "AD_ROSMAP_adj.P.Val")

#data[is.na(data)] <- 0
data <- na.omit(data)
data_goi <- subset(data, 
                        (AD_ROSMAP_logFC < -0.5 & AD_CWOW_logFC < -0.75) | 
                          (AD_ROSMAP_logFC > .5 & AD_CWOW_logFC > 0.5) | 
                          (AD_ROSMAP_logFC > .25 & AD_CWOW_logFC <= 0) |
                          (AD_ROSMAP_logFC <= 0 & AD_CWOW_logFC > 0.25))

#data$difference <- data$logFC - data$gene.log2FC
ggplot(data = data, 
  aes(x = AD_ROSMAP_logFC, y = AD_CWOW_logFC, text = paste(gene_name))) +
  annotate(
    "rect",
    xmin = 0,
    xmax = 1.75,
    ymin = 1.75,
    ymax = 0,
    fill = "lightpink3",
    alpha = .5
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -1.5,
    ymin = 0,
    ymax = -1.5,
    fill = "cadetblue3",
    alpha = .5) +
  geom_abline(color = "gray40") +
  geom_text_repel(
    data = data_goi, 
    aes(
      x = AD_ROSMAP_logFC, 
      y = AD_CWOW_logFC,
      label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 10)
  ) +
  geom_point(size = 1) + 
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
    title = "DEGs shared and unique between LBD CWOW AD and ROSMAP PCC",
    x = expression(paste("ROSMAP ", log[2](AD/control))),
    y = expression(paste("LBD CWOW ", log[2](AD/control))))+
  scale_y_continuous(breaks = seq(-1.5, 1.75, by = .5), 
                     limits = c(-1.5, 1.75), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-1.5, 1.75, by = .5), 
                     limits = c(-1.5, 1.75), expand = c(0,0))

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    condition, 
    "_AD_CWOW_vs_ROSMAP"
  )
#saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(data, x = "AD_ROSMAP_logFC", y = "AD_CWOW_logFC",
                add = "reg.line",  # Add regressin line
                add.params = list(color = "blue", fill = "lightgray"), 
                conf.int = TRUE)
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -.5, label.y = .75)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    condition, 
    "_AD_CWOW_vs_ROSMAP_corr_pval"
  )
#saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)
```

# CWOW vs ROSMAP - LBD 
```{r}
data <-
  merge(LBDvsControl,
        ROSMAP_AD,
        by = "gene_name",
        all = T)
data <- data[c(1,2,6,8,11,14,18,21,24)] #data[c(1,2,6,9,10,13,22,23,26)]
colnames(data) = c("gene_name", "chr", "gene_id", "LBD_CWOW_logFC", "LBD_CWOW_AveExpr", "LBD_CWOW_adj.P.Val", "AD_ROSMAP_logFC", "AD_ROSMAP_AveExpr", "AD_ROSMAP_adj.P.Val")

#data[is.na(data)] <- 0
data <- na.omit(data)
CWOW <- subset(data, LBD_CWOW_adj.P.Val <= 0.05)
ROSMAP <- subset(data, AD_ROSMAP_adj.P.Val <= 0.05) 

data <- subset(data, LBD_CWOW_adj.P.Val <= 0.05 & AD_ROSMAP_adj.P.Val <= 0.05)
data <- subset(data, LBD_CWOW_logFC <= -0.25 | LBD_CWOW_logFC >= 0.25)
data <- subset(data, AD_ROSMAP_logFC <= -0.25 | AD_ROSMAP_logFC >= 0.25)

data_goi <- subset(data, 
                        (AD_ROSMAP_logFC < -0.5 & LBD_CWOW_logFC < -0.75) | 
                          (AD_ROSMAP_logFC > .5 & LBD_CWOW_logFC > 0.5) | 
                          (AD_ROSMAP_logFC > .25 & LBD_CWOW_logFC <= 0) |
                          (AD_ROSMAP_logFC <= 0 & LBD_CWOW_logFC > 0.25))

#data$difference <- data$logFC - data$gene.log2FC
ggplot(data = data, 
  aes(x = AD_ROSMAP_logFC, y = LBD_CWOW_logFC, text = paste(gene_name))) +
  annotate(
    "rect",
    xmin = 0,
    xmax = 1.75,
    ymin = 1.75,
    ymax = 0,
    fill = "lightpink3",
    alpha = .5
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -1.5,
    ymin = 0,
    ymax = -1.5,
    fill = "cadetblue3",
    alpha = .5) +
  geom_abline(color = "gray40") +
  geom_text_repel(
    data = data_goi, 
    aes(
      x = AD_ROSMAP_logFC, 
      y = LBD_CWOW_logFC,
      label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 10)
  ) +
  geom_point(size = 1) + 
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
    title = "DEGs shared and unique between LBD CWOW and AD ROSMAP PCC",
    x = expression(paste("ROSMAP ", log[2](AD/control))),
    y = expression(paste("LBD CWOW ", log[2](AD/control))))+
  scale_y_continuous(breaks = seq(-1.5, 1.75, by = .5), 
                     limits = c(-1.5, 1.75), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-1.5, 1.75, by = .5), 
                     limits = c(-1.5, 1.75), expand = c(0,0))

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    condition, 
    "_LBD_CWOW_vs_AD_ROSMAP_FDRq0.05_log2FC0.25"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(data, x = "AD_ROSMAP_logFC", y = "LBD_CWOW_logFC",
                add = "reg.line",  # Add regressin line
                add.params = list(color = "blue", fill = "lightgray"), 
                conf.int = TRUE)
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -.5, label.y = .75)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    condition, 
    "_LBD_CWOW_vs_AD_ROSMAP_FDRq0.05_log2FC0.25_corr_pval"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)
```
```{r}
data <-
  merge(LBDvsControl,
        ROSMAP_AD,
        by = "gene_name",
        all = T)
data <- data[c(1,2,6,8,11,14,18,21,24)] #data[c(1,2,6,9,10,13,22,23,26)]
colnames(data) = c("gene_name", "chr", "gene_id", "LBD_CWOW_logFC", "LBD_CWOW_AveExpr", "LBD_CWOW_adj.P.Val", "AD_ROSMAP_logFC", "AD_ROSMAP_AveExpr", "AD_ROSMAP_adj.P.Val")

#data[is.na(data)] <- 0
data <- na.omit(data)
CWOW <- subset(data, LBD_CWOW_adj.P.Val <= 0.05)
ROSMAP <- subset(data, AD_ROSMAP_adj.P.Val <= 0.05) 

data <- subset(data, LBD_CWOW_adj.P.Val <= 0.05 & AD_ROSMAP_adj.P.Val <= 0.05)
data <- subset(data, LBD_CWOW_logFC <= -0.25 | LBD_CWOW_logFC >= 0.25)
data <- subset(data, AD_ROSMAP_logFC <= -0.25 | AD_ROSMAP_logFC >= 0.25)

data_goi <- subset(data, 
                        (AD_ROSMAP_logFC < -0.5 & LBD_CWOW_logFC < -0.75) | 
                          (AD_ROSMAP_logFC > .5 & LBD_CWOW_logFC > 0.5) | 
                          (AD_ROSMAP_logFC > .25 & LBD_CWOW_logFC <= 0) |
                          (AD_ROSMAP_logFC <= 0 & LBD_CWOW_logFC > 0.25))

#data$difference <- data$logFC - data$gene.log2FC
ggplot(data = data, 
  aes(x = AD_ROSMAP_logFC, y = LBD_CWOW_logFC, text = paste(gene_name))) +
  annotate(
    "rect",
    xmin = 0,
    xmax = 1.75,
    ymin = 1.75,
    ymax = 0,
    fill = "lightpink3",
    alpha = .5
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -1.5,
    ymin = 0,
    ymax = -1.5,
    fill = "cadetblue3",
    alpha = .5) +
  geom_abline(color = "gray40") +
  geom_text_repel(
    data = data_goi, 
    aes(
      x = AD_ROSMAP_logFC, 
      y = LBD_CWOW_logFC,
      label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 10)
  ) +
  geom_point(size = 1) + 
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
    title = "DEGs shared and unique between LBD CWOW and AD ROSMAP PCC",
    x = expression(paste("ROSMAP ", log[2](AD/control))),
    y = expression(paste("LBD CWOW ", log[2](AD/control))))+
  scale_y_continuous(breaks = seq(-1.5, 1.75, by = .5), 
                     limits = c(-1.5, 1.75), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-1.5, 1.75, by = .5), 
                     limits = c(-1.5, 1.75), expand = c(0,0))

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    condition, 
    "_LBD_CWOW_vs_AD_ROSMAP_FDRq0.05_log2FC0.25"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(data, x = "AD_ROSMAP_logFC", y = "LBD_CWOW_logFC",
                add = "reg.line",  # Add regressin line
                add.params = list(color = "blue", fill = "lightgray"), 
                conf.int = TRUE)
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -.5, label.y = .75)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    condition, 
    "_LBD_CWOW_vs_AD_ROSMAP_FDRq0.05_log2FC0.25_corr_pval"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

```
```{r}
data <-
  merge(LBDvsControl,
        ROSMAP_AD,
        by = "gene_name",
        all = T)
data <- data[c(1,2,6,8,11,14,18,21,24)] #data[c(1,2,6,9,10,13,22,23,26)]
colnames(data) = c("gene_name", "chr", "gene_id", "LBD_CWOW_logFC", "LBD_CWOW_AveExpr", "LBD_CWOW_adj.P.Val", "AD_ROSMAP_logFC", "AD_ROSMAP_AveExpr", "AD_ROSMAP_adj.P.Val")

#data[is.na(data)] <- 0
data <- na.omit(data)
CWOW <- subset(data, LBD_CWOW_adj.P.Val <= 0.05)
ROSMAP <- subset(data, AD_ROSMAP_adj.P.Val <= 0.05) 

data <- subset(data, LBD_CWOW_adj.P.Val <= 0.05 & AD_ROSMAP_adj.P.Val <= 0.05)
data <- subset(data, LBD_CWOW_logFC <= -0.25 | LBD_CWOW_logFC >= 0.25)
data <- subset(data, AD_ROSMAP_logFC <= -0.25 | AD_ROSMAP_logFC >= 0.25)

data_goi <- subset(data, 
                        (AD_ROSMAP_logFC < -0.5 & LBD_CWOW_logFC < -0.75) | 
                          (AD_ROSMAP_logFC > .5 & LBD_CWOW_logFC > 0.5) | 
                          (AD_ROSMAP_logFC > .25 & LBD_CWOW_logFC <= 0) |
                          (AD_ROSMAP_logFC <= 0 & LBD_CWOW_logFC > 0.25))

#data$difference <- data$logFC - data$gene.log2FC
ggplot(data = data, 
  aes(x = AD_ROSMAP_logFC, y = LBD_CWOW_logFC, text = paste(gene_name))) +
  annotate(
    "rect",
    xmin = 0,
    xmax = 1.75,
    ymin = 1.75,
    ymax = 0,
    fill = "lightpink3",
    alpha = .5
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -1.5,
    ymin = 0,
    ymax = -1.5,
    fill = "cadetblue3",
    alpha = .5) +
  geom_abline(color = "gray40") +
  geom_text_repel(
    data = data_goi, 
    aes(
      x = AD_ROSMAP_logFC, 
      y = LBD_CWOW_logFC,
      label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 10)
  ) +
  geom_point(size = 1) + 
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
    title = "DEGs shared and unique between LBD CWOW and AD ROSMAP PCC",
    x = expression(paste("ROSMAP ", log[2](AD/control))),
    y = expression(paste("LBD CWOW ", log[2](AD/control))))+
  scale_y_continuous(breaks = seq(-1.5, 1.75, by = .5), 
                     limits = c(-1.5, 1.75), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-1.5, 1.75, by = .5), 
                     limits = c(-1.5, 1.75), expand = c(0,0))

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    condition, 
    "_LBD_CWOW_vs_AD_ROSMAP_FDRq0.05_log2FC0.25"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(data, x = "AD_ROSMAP_logFC", y = "LBD_CWOW_logFC",
                add = "reg.line",  # Add regressin line
                add.params = list(color = "blue", fill = "lightgray"), 
                conf.int = TRUE)
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -.5, label.y = .75)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    condition, 
    "_LBD_CWOW_vs_AD_ROSMAP_FDRq0.05_log2FC0.25_corr_pval"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)
```

```{r}
data <-
  merge(LBDvsControl,
        ROSMAP_AD,
        by = "gene_name",
        all = T)
data <- data[c(1,2,6,8,11,14,18,21,24)] #data[c(1,2,6,9,10,13,22,23,26)]
colnames(data) = c("gene_name", "chr", "gene_id", "LBD_CWOW_logFC", "LBD_CWOW_AveExpr", "LBD_CWOW_adj.P.Val", "AD_ROSMAP_logFC", "AD_ROSMAP_AveExpr", "AD_ROSMAP_adj.P.Val")

#data[is.na(data)] <- 0
data <- na.omit(data)
data_goi <- subset(data, 
                        (AD_ROSMAP_logFC < -0.5 & LBD_CWOW_logFC < -0.75) | 
                          (AD_ROSMAP_logFC > .5 & LBD_CWOW_logFC > 0.5) | 
                          (AD_ROSMAP_logFC > .25 & LBD_CWOW_logFC <= 0) |
                          (AD_ROSMAP_logFC <= 0 & LBD_CWOW_logFC > 0.25))

#data$difference <- data$logFC - data$gene.log2FC
ggplot(data = data, 
  aes(x = AD_ROSMAP_logFC, y = LBD_CWOW_logFC, text = paste(gene_name))) +
  annotate(
    "rect",
    xmin = 0,
    xmax = 1.75,
    ymin = 1.75,
    ymax = 0,
    fill = "lightpink3",
    alpha = .5
  ) +  annotate(
    "rect",
    xmin = 0,
    xmax = -1.5,
    ymin = 0,
    ymax = -1.5,
    fill = "cadetblue3",
    alpha = .5) +
  geom_abline(color = "gray40") +
  geom_text_repel(
    data = data_goi, 
    aes(
      x = AD_ROSMAP_logFC, 
      y = LBD_CWOW_logFC,
      label = gene_name
    ),
    color = "black",
    fontface = "italic",
    size = 2.5, 
    max.overlaps = getOption("ggrepel.max.overlaps", default = 10)
  ) +
  geom_point(size = 1) + 
  theme_bw() +
  theme(plot.title = element_text(size = 10)) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_text(size = 10),
        axis.text.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10),
        axis.text.y = element_text(size = 10)) +
  labs(
    title = "DEGs shared and unique between LBD CWOW and AD ROSMAP PCC",
    x = expression(paste("ROSMAP ", log[2](AD/control))),
    y = expression(paste("LBD CWOW ", log[2](AD/control))))+
  scale_y_continuous(breaks = seq(-1.5, 1.75, by = .5), 
                     limits = c(-1.5, 1.75), expand = c(0,0)) +
  scale_x_continuous(breaks = seq(-1.5, 1.75, by = .5), 
                     limits = c(-1.5, 1.75), expand = c(0,0))

path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    condition, 
    "_LBD_CWOW_vs_AD_ROSMAP"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

# Scatter plot with correlation coefficient
sp <- ggscatter(data, x = "AD_ROSMAP_logFC", y = "LBD_CWOW_logFC",
                add = "reg.line",  # Add regressin line
                add.params = list(color = "blue", fill = "lightgray"), 
                conf.int = TRUE)
# Add correlation coefficient
sp + stat_cor(method = "pearson", label.x = -.5, label.y = .75)
path <-
  paste0(
    "../../results/",
    tool,
    "/correlation/",
    condition, 
    "_LBD_CWOW_vs_AD_ROSMAP_corr_pval"
  )
saveToPDF(paste0(path, ".pdf"), width = 6.25, height = 4.5)

```
```{r}
```
