==> scripts/R/09_WGNCA.Rmd <==
---
title: "WGNCA"
author: "Kimberly Olney, Ph.D."
date: "03/31/2023"
output:
  html_document:
    df_print: paged
  pdf_document: default
params:
  args: myarg
---
Tutorials: https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/Rpackages/WGCNA/Tutorials/index.html 
Background and glossary information https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/Rpackages/WGCNA/Tutorials/Simulated-00-Background.pdf 

Overview of steps:
1) data input
2) network construction
3) module detection
4) relating modules and genes to external information 

# Setup
```{r setup}
knitr::opts_knit$set(root.dir = ".")
```
# Libraries
GO enrichment calculations in WGCNA are deprecated recommended to use the R package anRichment which provides replacement for WGCNA functions GOenrichmentAnalysis() and userListEnrichment().
```{r install eval=FALSE}
BiocManager::install("WGCNA")
source("https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/GeneAnnotation/installAnRichment.R")
installAnRichment()
install.packages(c("AnnotationDBI", "GO.db", "org.Hs.eg.db", "org.Mm.eg.db", "TxDb.Hsapiens.UCSC.hg19.knownGene", "TxDb.Mmusculus.UCSC.mm10.knownGene"))
remotes::install_github("kevinblighe/CorLevelPlot")
```
```{r library}
library(WGCNA)
library(anRichmentMethods)
library(tidyverse)
library(CorLevelPlot)
library(gridExtra)
```
# User defined variables
```{r set_variables, warning=FALSE}
source(here::here("/research/labs/neurology/fryer/m239830/LBD_CWOW/scripts/R", "file_paths_and_colours.R"))
source(here::here("/research/labs/neurology/fryer/m239830/LBD_CWOW/scripts/R", "gtf_path.R"))
condition <- c("TYPE") #  TYPE
tool <- c("star")
```
# Read in DGE object & metadata
```{r dge}
dge.filtered.norm <- readRDS(paste0("../../rObjects/dge.filtered.norm.rds"))
# some samples are missing RIN values. 
# Replace NA with median RIN. 
# This is necessary to be able include RIN as a covariate in voom
# fill missing values with median
dge.filtered.norm$samples$RIN <- impute(dge.filtered.norm$samples$RIN, median)
# one sample is missing VaD information
dge.filtered.norm$samples$VaD <- impute(dge.filtered.norm$samples$VaD, median)
dge.filtered.norm$samples$flowcell_and_lane <- factor(dge.filtered.norm$samples$flowcell_and_lane)
dge.filtered.norm$samples$APOE <- factor(dge.filtered.norm$samples$APOE)

info <- as.data.frame(dge.filtered.norm$samples)
genes <- dge.filtered.norm$genes
```

# Add biomarker expression to dge metadata 
```{r genes_lcpm}
genes <- dge.filtered.norm$genes # gene information 
lcpm <- edgeR::cpm(dge.filtered.norm$counts, log = TRUE) # obtain log CPM counts
```

```{r biomarker_expression}
biomarkers <- c("ENO2", "GFAP", "OLIG2", "CD34", "P2RY12")

for (i in biomarkers) {
  biomarker <- subset(genes, gene_name == i) # gene of interest 
  biomarker_counts <- subset(lcpm, rownames(lcpm) %in% biomarker)
  biomarker_melt <- reshape2::melt(biomarker_counts) # reshape data 
  # rename columns to merge with metadata 
  names(biomarker_melt)[names(biomarker_melt) == "value"] <- i 
  # rename columns to merge with metadata 
  names(biomarker_melt)[names(biomarker_melt) == "Var2"] <- "NPID" 
  names(biomarker_melt)[names(biomarker_melt) == "Var1"] <- "gene_id"
  biomarker_melt$gene_id <- NULL
  assign(paste0(i),biomarker_melt)
}

# add gene expression values into one dataframe 
# put all data frames into list
df_list <- list(ENO2, OLIG2, CD34, P2RY12, GFAP)

# merge all data frames in list
cell_biomarker_lcpm <- df_list %>% reduce(full_join, by='NPID')
```
# Scale data
rescaling a predictor in a regression has absolutely no effect on the magnitude of the relation being studied—the slope itself will not change its steepness, nor will the p-values or variance explained be changed. Rescaling is merely a means of communicating the nature of the regression line in different, hopefully more intuitive language.
```{r scale}
df <- merge(cell_biomarker_lcpm, info, by = "NPID")
scaled.info <-
  df[c(
    "Race_numeric",
    "RIN",
    "Age",
    "PCT_CODING_BASES",
    "PCT_INTERGENIC_BASES",
    "PCT_INTRONIC_BASES",
    "APOE_E4_allele_count", 
    "ENO2", 
    "GFAP", 
    "OLIG2", 
    "CD34", 
    "P2RY12",
    "Brain.wt"
  )] %>% scale()
scaled.info.df <- as.data.frame(scaled.info)
# remove columns with unscaled data 
df <- (subset(df, select = -c(Race_numeric, RIN, Age, PCT_CODING_BASES, PCT_INTRONIC_BASES, APOE_E4_allele_count, ENO2, GFAP, OLIG2, CD34, P2RY12, Brain.wt)))
# Add scaled information to the metadata called "info"
info_with_scale <- cbind(df, scaled.info.df)
```

```{r}
all.equal(dge.filtered.norm$samples$NPID, as.character(info_with_scale$NPID))
# replace sample information with the updated info that includes biomakrer expression
dge.filtered.norm$samples <- info_with_scale
```
Voom transform counts to use for BIC 
```{r voom, warning=FALSE}
formula <- (~ 0 + TYPE)
voom_with_weights <-
  variancePartition::voomWithDreamWeights(
    counts = dge.filtered.norm$counts,
    formula = formula,
    data = dge.filtered.norm$samples,
    BPPARAM = BiocParallel::SnowParam(cores),
    plot = TRUE
  )
path <- paste0("../../results/", tool, "/voom/", condition, ".raw.voom")
saveToPDF(paste0(path, ".pdf"), width = 6, height = 4)
voomCounts <- voom_with_weights$E
```
# User defined variables
```{r set_variables}
source(here::here("scripts/R", "file_paths_and_colours.R"))
condition <- c("TYPE")
tool <- "star"
```
# Data traits
```{r}
counts <- t(voomCounts)
dim(counts)
dim(info)
# remove columns that hold information we do not need.
#dataTraits <- info[, -c(1:3,5:10,20,22,24:27,29,31:39,41:123)] # remove
dataTraits <- info[, c(6,11:19,21,22,24,29,37,108,109,114,117:124)] # Keep
dim(dataTraits)
# disease type isn't numeric, make numeric as this is required down stream
numeric_traits <- sapply(dataTraits, as.numeric)
numeric_traits <- as.data.frame(numeric_traits) # save as data frame
rownames(numeric_traits) <- rownames(dataTraits) # add NPID as row names
```
# check data
```{r}
all.equal(info$NPID, rownames(counts))
all.equal(info$NPID, colnames(lcpm))
```
# QC - outlier detection
```{r}
gsg <- goodSamplesGenes(counts)
summary(gsg)

table(gsg$goodGenes) # outlier genes
table(gsg$goodSamples) # outlier samples
# no outliers
```
# hclust & pca
```{r}
htree <- hclust(dist(counts))
plot(htree)

pca <- prcomp(counts)
pca.dat <- pca$x
pca.var <- pca$sdev^2
pca.var.percent <- round(pca.var/sum(pca.var)*100, digits = 2)
pca.dat <- as.data.frame(pca.dat)
ggplot(pca.dat, aes(PC1, PC2)) +
  geom_point() +
  geom_text(label = rownames(pca.dat)) +
  labs(x = paste0('PC1: ', pca.var.percent[1], '%'))
```
# Correct for batch effects 
```{r}
design <-
  model.matrix(~ 0 + 
      TYPE,
    dge.filtered.norm$samples
  )

colnames(design) <-
  c(
    CONTROL,
    PA,
    AD,
    LBD
  )
rownames(design) <- dge.filtered.norm$samples$NPID

covs <-
  model.matrix(~ 0 +
      sex_numeric + 
      RIN +
      PCT_CODING_BASES +
      PCT_INTERGENIC_BASES + 
      PCT_INTRONIC_BASES + 
      ENO2, 
    dge.filtered.norm$samples
  )

colnames(covs) <-
  c(
    "sex",
    "RIN",
    "PCT_CODING_BASES", 
    "PCT_INTERGENIC_BASES", 
    "PCT_INTRONIC_BASES",
    "ENO2"
  )
rownames(covs) <- dge.filtered.norm$samples$NPID
dim(covs)
anew <- removeBatchEffect(lcpm, batch=dge.filtered.norm$samples$flowcell_and_lane, covariates = covs, design=design)
anew <- t(anew)
```
# hclust & pca with batch corrected counts
```{r}
htree <- hclust(dist(anew))
plot(htree)

pca <- prcomp(anew)
pca.dat <- pca$x
pca.var <- pca$sdev^2
pca.var.percent <- round(pca.var/sum(pca.var)*100, digits = 2)
pca.dat <- as.data.frame(pca.dat)
ggplot(pca.dat, aes(PC1, PC2)) +
  geom_point() +
  geom_text(label = rownames(pca.dat)) +
  labs(x = paste0('PC1: ', pca.var.percent[1], '%'))

gsg <- goodSamplesGenes(anew)
summary(gsg)

table(gsg$goodGenes) # outlier genes - outlier is ENO2
table(gsg$goodSamples) # outlier samples
# no outliers
```
# All equal 
```{r}
all(rownames(info) == rownames(counts))
```

# Re-format & sample cluster
Next we cluster the samples (in contrast to clustering genes that will come later) to see if there are any obvious outliers.

Hierarchical cluster analysis and the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).
### complete linkage
```{r cluster_complete_linkage}
# Cluster samples by Hierarchical cluster analysis 
# The hclust function in R uses the complete linkage method for hierarchical clustering by default. 
# This particular clustering method defines the cluster distance between two clusters to be the maximum distance between their individual components.
sampleTree <- hclust(dist(counts))
# Plot the sample tree
par(cex = 1) # point size 
par(mar = c(0, 4, 1, 0)) # plot marigns  bottom, left, top and right 
plot(
  sampleTree,
  main = "Sample clustering to detect outliers",
  sub = "",
  xlab = "",
  cex.lab = 1.5,
  cex.axis = 1.5,
  cex.main = 2
)
path <- paste0("../../results/star/WGNCA/sample_clusters_voomcounts_method_complete_linkage")
saveToPDF(paste0(path, ".pdf"), width = 60, height = 20)

remove(sampleTree)
```
### ward.D2
```{r cluster_ward}
# Cluster samples by Hierarchical cluster analysis 
# single The single linkage method (which is closely related to the minimal spanning tree)
# adopts a ‘friends of friends’ clustering strategy.
sampleTree <- hclust(dist(counts), method = "ward.D2")
# Plot the sample tree
par(cex = 1) # point size 
par(mar = c(0, 4, 1, 0)) # plot marigns  bottom, left, top and right 
plot(
  sampleTree,
  main = "Sample clustering to detect outliers",
  sub = "",
  xlab = "",
  cex.lab = 1.5,
  cex.axis = 1.5,
  cex.main = 2
)
path <- paste0("../../results/star/WGNCA/sample_clusters_voomcounts_method_ward.D2")
saveToPDF(paste0(path, ".pdf"), width = 60, height = 20)

remove(sampleTree)
```
### single 
```{r cluster_single}
# Cluster samples by Hierarchical cluster analysis 
# single The single linkage method (which is closely related to the minimal spanning tree)
# adopts a ‘friends of friends’ clustering strategy.
sampleTree <- hclust(dist(counts), method = "single")
# Plot the sample tree
par(cex = 1) # point size 
par(mar = c(0, 4, 1, 0)) # plot marigns  bottom, left, top and right 
plot(
  sampleTree,
  main = "Sample clustering to detect outliers",
  sub = "",
  xlab = "",
  cex.lab = 1.5,
  cex.axis = 1.5,
  cex.main = 2
)
path <- paste0("../../results/star/WGNCA/sample_clusters_voomcounts_method_single")
saveToPDF(paste0(path, ".pdf"), width = 60, height = 20)

remove(sampleTree)
```
### average 
```{r cluster_average}
sampleTree <- hclust(dist(counts))
# Plot the sample tree
par(cex = 1) # point size 
par(mar = c(0, 4, 1, 0)) # plot marigns  bottom, left, top and right 
plot(
  sampleTree,
  main = "Sample clustering to detect outliers",
  sub = "",
  xlab = "",
  cex.lab = 1.5,
  cex.axis = 1.5,
  cex.main = 2
)
path <- paste0("../../results/star/WGNCA/sample_clusters_voomcounts_method_average")
saveToPDF(paste0(path, ".pdf"), width = 60, height = 20)
```

# Remove outliers
Choose a height cut that will remove the offending samples, say 6000 (the red line in the plot), and use a branch cut at
that height.
```{r eval = FALSE}
plot(
  sampleTree,
  main = "Sample clustering to detect outliers",
  sub = "",
  xlab = "",
  cex.lab = 1.5,
  cex.axis = 1.5,
  cex.main = 2
)
# Plot a line to show the cut
abline(h = 175, col = "red")
# Determine cluster under the line
# cutHeight = height at which branches are to be cut.
# minSize = minimum number of object on a branch to be considered a cluster.
clust <- cutreeStatic(sampleTree, cutHeight = 175, minSize = 10)
table(clust)
# clust 1 contains the samples we want to keep.
keepSamples <- (clust == 1)
datExpr <- counts[keepSamples,] # keep samples that are not outliers 
nGenes <- ncol(datExpr) # number of genes
nSamples <- nrow(datExpr) # number of samples 
```

# Outliers removed 
```{r}
# Form a data frame analogous to expression data that will hold the clinical traits.
NPID <- rownames(counts)
traitRows <- match(NPID, rownames(numeric_traits))
numeric_traits <- numeric_traits[traitRows, ]
collectGarbage() # Performs garbage collection until free memory idicators show no change.
```
# Sample dendrogram & trait heatmap
```{r}
# Re-cluster samples
sampleTree2 <- hclust(dist(counts), method = "average")

# Convert traits to a color representation:
# white means low, red means high, grey means missing entry
traitColors <- numbers2colors(numeric_traits, signed = FALSE)
# Plot the sample dendrogram and the colors underneath.
plotDendroAndColors(sampleTree2,
                    traitColors,
                    groupLabels = names(numeric_traits),
                    main = "Sample dendrogram and trait heatmap")
path <- paste0("../../results/star/WGNCA/sample_dendrogram_and_trait_heatmap")
saveToPDF(paste0(path, ".pdf"), width = 60, height = 20)
```
# Save
```{r}
save(voomcounts, dataTraits, file = "allsamplesDataInput.RData")
```
# Automatic block-wise network construction and module detection 
```{r}
# load the data
lnames = load(file = "allsamplesDataInput.RData");
#The variable lnames contains the names of loaded variables.
lnames
data_counts <- sapply(voomcounts, as.numeric)
gsg = goodSamplesGenes(counts, verbose = 3)
gsg$allOK
```
# Power
```{r}

# Choose a soft threshold power- USE A SUPERCOMPUTER IRL ------------------------------------
  
powers = c(c(1:10), seq(from =10, to=30, by=1)) #choosing a set of soft-thresholding powers
sft = pickSoftThreshold(counts, powerVector=powers, verbose =5, networkType="signed") #call network topology analysis function
  
sizeGrWindow(9,5)
par(mfrow= c(1,2))
cex1=0.9
plot(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2], xlab= "Soft Threshold (power)", ylab="Scale Free Topology Model Fit, signed R^2", type= "n", main= paste("Scale independence"))
text(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2], labels=powers, cex=cex1, col="red")
abline(h=0.90, col="red")
plot(sft$fitIndices[,1], sft$fitIndices[,5], xlab= "Soft Threshold (power)", ylab="Mean Connectivity", type="n", main = paste("Mean connectivity"))
text(sft$fitIndices[,1], sft$fitIndices[,5], labels=powers, cex=cex1, col="red")
 
#from this plot, we would choose a power of 12 because it's the lowest power for which the scale free topology index reaches 0.80
```
# Construct a gene co-expression matrix and generate modules
```{r}
#build a adjacency "correlation" matrix
softPower = 12
adjacency = adjacency(counts, power = softPower, type = "signed") #specify network type
head(adjacency)
 
# Construct Networks- USE A SUPERCOMPUTER IRL -----------------------------
#translate the adjacency into topological overlap matrix and calculate the corresponding dissimilarity:
TOM = TOMsimilarity(adjacency, TOMType="signed") # specify network type
dissTOM = 1-TOM
 
# Generate Modules --------------------------------------------------------
install.packages("flashClust")
library("flashClust")
# Generate a clustered gene tree
geneTree = flashClust(as.dist(dissTOM), method="average")
plot(geneTree, xlab="", sub="", main= "Gene Clustering on TOM-based dissimilarity", labels= FALSE, hang=0.04)
#This sets the minimum number of genes to cluster into a module
minModuleSize = 30
dynamicMods = cutreeDynamic(dendro= geneTree, distM= dissTOM, deepSplit=2, pamRespectsDendro= FALSE, minClusterSize = minModuleSize)
dynamicColors= labels2colors(dynamicMods)
MEList= moduleEigengenes(counts, colors= dynamicColors,softPower = softPower)
MEs= MEList$eigengenes
MEDiss= 1-cor(MEs)
METree= flashClust(as.dist(MEDiss), method= "average")
save(dynamicMods, MEList, MEs, MEDiss, METree, file= "Network_allSamples_signed_RLDfiltered.RData")
 
 
#plots tree showing how the eigengenes cluster together
#INCLUE THE NEXT LINE TO SAVE TO FILE
#pdf(file="clusterwithoutmodulecolors.pdf")
plot(METree, main= "Clustering of module eigengenes", xlab= "", sub= "")
#set a threhold for merging modules. In this example we are not merging so MEDissThres=0.0
MEDissThres = 0.0
merge = mergeCloseModules(counts, dynamicColors, cutHeight= MEDissThres, verbose =3)
mergedColors = merge$colors
mergedMEs = merge$newMEs
#INCLUE THE NEXT LINE TO SAVE TO FILE
#dev.off()
 
#plot dendrogram with module colors below it
#INCLUE THE NEXT LINE TO SAVE TO FILE
#pdf(file="cluster.pdf")
plotDendroAndColors(geneTree, cbind(dynamicColors, mergedColors), c("Dynamic Tree Cut", "Merged dynamic"), dendroLabels= FALSE, hang=0.03, addGuide= TRUE, guideHang=0.05)
moduleColors = mergedColors
colorOrder = c("grey", standardColors(50))
moduleLabels = match(moduleColors, colorOrder)-1
MEs = mergedMEs
#INCLUE THE NEXT LINE TO SAVE TO FILE
#dev.off()
 
save(MEs, moduleLabels, moduleColors, geneTree, file= "Network_allSamples_signed_nomerge_RLDfiltered.RData")

names(counts)[moduleColors=="grey60"]
```
```{r}
# Correlate traits --------------------------------------------------------
 
 
#Define number of genes and samples
nGenes = ncol(counts)
nSamples = nrow(counts)
#Recalculate MEs with color labels
MEs0 = moduleEigengenes(counts, moduleColors)$eigengenes
MEs = orderMEs(MEs0)
moduleTraitCor = cor(MEs, numeric_traits, use= "p")
moduleTraitPvalue = corPvalueStudent(moduleTraitCor, nSamples)
 
 
#Print correlation heatmap between modules and traits
textMatrix= paste(signif(moduleTraitCor, 2), "\n(",
                        signif(moduleTraitPvalue, 1), ")", sep= "")
dim(textMatrix)= dim(moduleTraitCor)
par(mar= c(6, 8.5, 3, 3))
 
 
#display the corelation values with a heatmap plot
#INCLUE THE NEXT LINE TO SAVE TO FILE
pdf(file="heatmap.pdf", width = 20, height = 15)
labeledHeatmap(Matrix= moduleTraitCor,
            xLabels= names(numeric_traits),
            yLabels= names(MEs),
            ySymbols= names(MEs),
            colorLabels= FALSE,
            colors= blueWhiteRed(50),
            textMatrix= textMatrix,
            setStdMargins= FALSE,
            cex.text= 0.5,
            zlim= c(-1,1),
            main= paste("Module-trait relationships"))
#INCLUE THE NEXT LINE TO SAVE TO FILE
dev.off()
```
```{r}
datKME_WT=signedKME(counts, mergedMEs, outputColumnName="MM.")
keeps <- c("MM.tan")
df_tan = datKME_WT[keeps]
tan <- subset(df_tan, df_tan$MM.tan > 0.8)
tan_genes <- rownames(tan)
write.table(tan_genes, "tan_genes.txt", quote = FALSE, row.names = FALSE)
names(datExpr) -> names(mergedColors_WT)
matrix_mergedColors_WT <-as.matrix(mergedColors_WT)
datKME_moduleColor_WT<-merge(matrix_mergedColors_WT, datKME_WT, by =  "row.names")
head(datKME_moduleColor_WT)
names(datKME_WT)[moduleColors=="tan"]
```

```{r}
# Choose a set of soft-thresholding powers
powers = c(c(1:10), seq(from = 12, to=20, by=2))
# Call the network topology analysis function
sft = pickSoftThreshold(tcounts, powerVector = powers, verbose = 5)
# Plot the results:
sizeGrWindow(9, 5)
par(mfrow = c(1,2));
cex1 = 0.9;
# Scale-free topology fit index as a function of the soft-thresholding power
plot(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],
     xlab="Soft Threshold (power)",ylab="Scale Free Topology Model Fit,signed R^2",type="n",
     main = paste("Scale independence"));
text(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],
     labels=powers,cex=cex1,col="red"); 
# this line corresponds to using an R^2 cut-off of h
abline(h=0.90,col="red")
# Mean connectivity as a function of the soft-thresholding power
plot(sft$fitIndices[,1], sft$fitIndices[,5],
     xlab="Soft Threshold (power)",ylab="Mean Connectivity", type="n",
     main = paste("Mean connectivity"))
text(sft$fitIndices[,1], sft$fitIndices[,5], labels=powers, cex=cex1,col="red")
```

# blockwise
```{r}
dim(data_counts)
bwnet = blockwiseModules(voomcounts, maxBlockSize = 15000,
                     power = 14,
                     verbose = 3)
```
bwnet$colors contains the module assignment, and bwnet$MEs contains the module eigengenes of the modules.
```{r}
# Load the results of single-block analysis
save(bwnet, file = "allsamples_bwnet.RData")
```

```{r}
# open a graphics window
sizeGrWindow(12, 9)
# Convert labels to colors for plotting
mergedColors = labels2colors(bwnet$colors)
# Plot the dendrogram and the module colors underneath
plotDendroAndColors(bwnet$dendrograms[[1]], mergedColors[bwnet$blockGenes[[1]]],
"Module colors",
dendroLabels = FALSE, hang = 0.03,
addGuide = TRUE, guideHang = 0.05)

moduleLabels = bwnet$colors
moduleColors = labels2colors(bwnet$colors)
MEs = bwnet$MEs;
geneTree = bwnet$dendrograms[[1]];
save(MEs, moduleLabels, moduleColors, geneTree,
file = "02-networkConstruction-auto.RData")

#load(file = "FemaleLiver-02-networkConstruction-auto.RData");
# Relabel blockwise modules
bwLabels = matchLabels(bwnet$colors, moduleLabels)
# Convert labels to colors for plotting
bwModuleColors = labels2colors(bwLabels)
table(bwLabels)
```
```{r}
# open a graphics window
#sizeGrWindow(6,6)
# Plot the dendrogram and the module colors underneath for block 1
plotDendroAndColors(bwnet$dendrograms[[1]], bwModuleColors[bwnet$blockGenes[[1]]],
                  "Module colors", main = "Gene dendrogram and module colors in block 1",
                  dendroLabels = FALSE, hang = 0.03,
                  addGuide = TRUE, guideHang = 0.05)

```
```{r}
moduleLabels = bwnet$colors
moduleColors = labels2colors(bwnet$colors)
MEs = bwnet$MEs;
geneTree = bwnet$dendrograms[[1]];
save(MEs, moduleLabels, moduleColors, geneTree,
file = "02-networkConstruction-auto.RData")
```
# Relate modules to variables 
```{r}
#Load the expression and trait data saved in the first part
lnames = load(file = "allsamplesDataInput.RData");
#The variable lnames contains the names of loaded variables.
lnames
# Load network data saved in the second part.
lnames = load(file = "02-networkConstruction-auto.RData");
lnames
```
Quantifying module–trait associations
```{r}
tcounts <- sapply(tcounts, as.numeric)
# Define numbers of genes and samples
nGenes = ncol(tcounts)
nSamples = nrow(tcounts)
# Recalculate MEs with color labels
MEs0 = moduleEigengenes(tcounts, moduleColors)$eigengenes
MEs = orderMEs(MEs0)
moduleTraitCor = cor(MEs, dataTraits, use = "p")
moduleTraitPvalue = corPvalueStudent(moduleTraitCor, nSamples)
```

```{r}
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
                        signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)
par(mar = c(9, 8.5, 1, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = moduleTraitCor,
             xLabels = names(dataTraits),
             yLabels = names(MEs),
             ySymbols = names(MEs),
             colorLabels = FALSE,
             colors = blueWhiteRed(50),
             textMatrix = textMatrix,
             setStdMargins = FALSE,
             cex.text = 0.5,
             zlim = c(-1,1),
             main = paste("Module-trait relationships"))

path <- paste0("../../results/star/WGNCA/module-trait_relationships")
saveToPDF(paste0(path, ".pdf"), width = 12, height = 7)
```
